% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem_view.R
\name{annotatePubchemCompound}
\alias{annotatePubchemCompound}
\title{Annotate PubChem Compound}
\usage{
annotatePubchemCompound(
  cid,
  heading = "ChEMBL ID",
  source = NULL,
  parse_function = identity
)
}
\arguments{
\item{cid}{The compound ID (CID) of the PubChem compound.}

\item{heading}{The type of information to retrieve. Default is "ChEMBL ID".}

\item{source}{The data source to use. Default is NULL.}

\item{parse_function}{A custom parsing function to process the response. Default is the identity function.}

\item{output}{The format of the output. Default is "JSON". Options include "JSON", "XML", "ASN"}
}
\value{
The annotated information about the PubChem compound.
}
\description{
This function retrieves information about a PubChem compound based on the provided compound ID (CID).
}
\examples{
annotatePubchemCompound(cid = 12345)
annotatePubchemCompound(cid = 67890, heading = "CAS")

}
