% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationGx-package.R
\docType{package}
\name{AnnotationGx-package}
\alias{AnnotationGx}
\alias{AnnotationGx-package}
\title{AnnotationGx: AnnotationGx: A package for building, updating and querying an annotation database for pharmaco-genomic data}
\description{
A package for building, updating and querying an annotation database for pharmaco-genomic data. Provides API interfaces to various bioinformatic and chemoinformatic databases, as well as a set of functions for curating your data using the annotations retrieved from those APIs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bhklab.github.io/AnnotationGx/}
  \item Report bugs at \url{https://github.com/bhklab/AnnotationGx/issues/new?template=issue_template.md}
}

}
\author{
\strong{Maintainer}: Benjamin Haibe-Kains \email{benjamin.haibe.kains@utoronto.ca}

Authors:
\itemize{
  \item Benjamin Haibe-Kains \email{benjamin.haibe.kains@utoronto.ca}
  \item Michael Tran \email{michaelcao-anh.tran@uhn.ca}
  \item Jermiah Joseph \email{jermiah.joseph@gmail.com}
  \item Christopher Eeles \email{christopher.eeles@uhnresearch.ca}
}

}
\keyword{internal}
