% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-API.R
\docType{class}
\name{API-class}
\alias{API-class}
\alias{.API}
\title{An API Connection Class}
\description{
A class representing a connection to an API at a specific URL. Stores
and automatically retrieves a set of metadata associated with the API
including any documentation, if available, the types of requests it accepts
and what targets it exposes for queries.
}
\section{Slots}{

\describe{
\item{\code{url}}{\code{character}}

\item{\code{documentation}}{\code{list}}

\item{\code{accepts}}{\code{character}}

\item{\code{targets}}{\code{character}}
}}

\keyword{internal}
