% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-API.R
\name{API}
\alias{API}
\title{Constructor for an API Connection Class}
\usage{
API(url)
}
\arguments{
\item{url}{\code{character} the url of the API you wish to connect to}
}
\value{
an \code{API} object storing the API connection and metadata about it.
}
\description{
Constructor for an API Connection Class
}
