% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_core.R
\name{bm_query_builder}
\alias{bm_query_builder}
\title{Build BioMart Query XML}
\usage{
bm_query_builder(
  dataset,
  filters = list(),
  attributes = character(),
  client_name = "biomartclient",
  processor = "TSV",
  header = TRUE,
  limit = -1
)
}
\arguments{
\item{dataset}{A DatasetInfo object representing the BioMart dataset to query}

\item{filters}{List of filters, either:
\itemize{
\item Named list where names are filter names and values are filter values
\item List of FilterInfo objects with added 'value' field
}}

\item{attributes}{AttributeSet, list of AttributeInfo objects, or character vector
specifying the attributes to retrieve}

\item{client_name}{Character, name of client for query identification (default: "biomartclient")}

\item{processor}{Character, output format processor type (default: "TSV")}

\item{header}{Logical, whether to include header row in results (default: TRUE)}

\item{limit}{Integer, maximum number of rows to return, -1 for unlimited (default: -1)}
}
\value{
Character string containing the formatted XML BioMart query
}
\description{
Build BioMart Query XML
}
