% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem_rest.R
\name{build_pubchem_rest_query}
\alias{build_pubchem_rest_query}
\title{Query PubChem REST API}
\usage{
build_pubchem_rest_query(
  id,
  domain = "compound",
  namespace = "name",
  operation = "cids",
  output = "JSON",
  url = "https://pubchem.ncbi.nlm.nih.gov/rest/pug",
  raw = FALSE,
  query_only = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{id}{The identifier(s) of the entity to query. For compound queries, it can be a single value or a vector of values. For other domains, it must be a single value.}

\item{domain}{The domain to query. Options are 'compound', 'substance', 'assay', 'cell', 'gene', and 'protein'.}

\item{namespace}{The namespace to use for the query. The available options depend on the chosen domain.}

\item{operation}{The operation to perform. The available options depend on the chosen domain and namespace.}

\item{output}{The desired output format. Options are 'JSON', 'XML', 'SDF', 'TXT', and 'CSV'.}

\item{url}{The base URL of the PubChem REST API.}

\item{raw}{Logical indicating whether to return the raw response or parse it into a structured format.}

\item{query_only}{Logical indicating whether to return the constructed query URL without making the actual request.}

\item{verbose}{Logical indicating whether to print debug information.}

\item{...}{Additional arguments to be passed to the underlying HTTP request function.}
}
\value{
The response from the PubChem REST API, parsed into a structured format if raw is set to FALSE.
}
\description{
This function queries the PubChem REST API to retrieve information
about compounds, substances, assays, cells, genes, and proteins.
}
\references{
PubChem REST API documentation: https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest
}
