% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChem.R
\name{buildAIDTable}
\alias{buildAIDTable}
\title{Build a \code{data.table} of assay ids from the a PubChem query list.}
\usage{
buildAIDTable(list)
}
\description{
Build a \code{data.table} of assay ids from the a PubChem query list.
}
