% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_model_passports_models}
\alias{cell_model_passports_models}
\title{cell_model_passports_models is a preprocessed version of the cell model passports dataset}
\format{
A data table with 8 columns and 1001 rows.
\describe{
\item{CMP.model_id}{\code{char} The model id of the cell line in the cell model passports dataset.}
\item{CMP.sample_id}{\code{char} The sample id of the cell line in the cell model passports dataset.}
\item{CMP.model_name}{\code{char} The name of the cell line in the cell model passports dataset.}
\item{CMP.cancer_type_ncit_id}{\code{char} The cancer type ncit id of the cell line in the cell model passports dataset.}
\item{CMP.COSMIC_ID}{\code{int} The COSMIC ID of the cell line in the cell model passports dataset.}
\item{CMP.BROAD_ID}{\code{int} The BROAD ID of the cell line in the cell model passports dataset.}
\item{CMP.CCLE_ID}{\code{int} The CCLE ID of the cell line in the cell model passports dataset.}
\item{CMP.RRID}{\code{char} The RRID of the cell line in the cell model passports dataset.}
}
}
\source{
https://cog.sanger.ac.uk/cmp/download/model_list_20240103.csv
}
\usage{
data(cell_model_passports_models)
}
\description{
A preprocessed version of the cell model passports dataset. This dataset
contains the following columns:
CMP.model_id, CMP.sampleid, CMP.model_name, CMP.cancer_type_ncit_id,
CMP.COSMIC_ID, CMP.BROAD_ID, CMP.CCLE_ID, and CMP.RRID.
This dataset is used in the AnnotationGx package to map cell line names from
various sources to the Cellosaurus database.
}
\examples{
data(cell_model_passports_models)
head(cell_model_passports_models)
}
\keyword{datasets}
