% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{cellosaurus_fields}
\alias{cellosaurus_fields}
\title{Get the list of fields in the Cellosaurus schema}
\usage{
cellosaurus_fields(common = FALSE, upper = FALSE)
}
\arguments{
\item{common}{Logical indicating whether to return only the common fields. Default is FALSE.}

\item{upper}{Logical indicating whether to return the fields in uppercase. Default is FALSE.}
}
\value{
A character vector containing the list of fields in the Cellosaurus schema.
}
\description{
This function retrieves the list of fields available in the Cellosaurus schema.
It internally calls the \code{.cellosaurus_schema()} function to fetch the schema
and extracts the list of fields from it.
}
\examples{
cellosaurus_fields()
cellosaurus_fields(common = TRUE)
cellosaurus_fields(upper = TRUE)

}
