% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{cellosaurusAPIVersion}
\alias{cellosaurusAPIVersion}
\title{Get the Cellosaurus API version}
\usage{
cellosaurusAPIVersion()
}
\value{
The version of the Cellosaurus API.
}
\description{
This function retrieves the version of the Cellosaurus API.
}
\examples{
cellosaurusAPIVersion()

}
