% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{characterToNamedVector}
\alias{characterToNamedVector}
\title{Convert a character string to a named vector.}
\usage{
characterToNamedVector(x)
}
\arguments{
\item{x}{A character string in the format "name=value|name=value|..."}
}
\value{
A named vector with the names and values extracted from the input string.
}
\description{
This function takes a character string in the format "name=value|name=value|..."
and converts it into a named vector, where each name-value pair is separated by
a pipe symbol (|) and the name and value are separated by an equals sign (=).
}
