% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellosaurusAPI.R
\name{cleanCellosaurusResponse}
\alias{cleanCellosaurusResponse}
\title{Clean cellosaurus responses}
\usage{
cleanCellosaurusResponse(
  responseList,
  GETfxn = c("search/cell-line?", "cell-line/")
)
}
\arguments{
\item{responseList}{is a list of responses}
}
\value{
A list of responses
}
\description{
This function takes a list of Cellosaurus Responses and cleans them for use
}
\details{
This function takes a list of Cellosaurus Responses and cleans them for use
}
