% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-standardize_names.R
\name{cleanCharacterStrings}
\alias{cleanCharacterStrings}
\title{Clean character strings by removing special characters and formatting.}
\usage{
cleanCharacterStrings(name, space_action = "")
}
\arguments{
\item{name}{A character string to be cleaned.}

\item{space_action}{A character vector specifying the actions to be taken for space characters.
One of c("", "-", " ").}
}
\value{
The cleaned character string.
}
\description{
This function takes a character string as input and performs several cleaning operations
to remove special characters, formatting, and unwanted substrings. The cleaned string
is then returned as the output.
}
\examples{
cleanCharacterStrings("Cisplatin: 1 mg/mL (1.5 mM); 5 mM in DMSO")

}
