% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChEMBL.R
\name{compoundQuery}
\alias{compoundQuery}
\title{Query the ChEMBL compound API end-point}
\usage{
compoundQuery(field, filter_type, value)
}
\arguments{
\item{field}{The field to query.}

\item{filter_type}{The type of filter to apply.}

\item{value}{The value to filter on.}
}
\value{
A response object containing the query results.
}
\description{
This function is used to query the compound_record resource from the ChEMBL API.
It takes three parameters: field, filter_type, and value.
The field parameter specifies the field to query, such as compound_key, compound_name, document_chembl_id, record_id, or src_id.
The filter_type parameter specifies the type of filter to apply, such as "exact", "contains", "startswith", or "endswith".
The value parameter specifies the value to filter on.
}
