% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChEMBL.R
\name{constructChemblQuery}
\alias{constructChemblQuery}
\title{A general function for creating Queries to the ChEMBL API}
\usage{
constructChemblQuery(resource, field, filter_type, value)
}
\description{
A general function for creating Queries to the ChEMBL API\tabular{cc}{
   Resource Name \tab Description \cr
   activity \tab Activity values recorded in an Assay \cr
   assay \tab Assay details as reported in source Document/Dataset \cr
   atc_class \tab WHO ATC Classification for drugs \cr
   binding_site \tab WHO ATC Classification for drugs \cr
   biotherapeutic \tab Biotherapeutic molecules, which includes HELM notation and sequence data \cr
   cell_line \tab Cell line information \cr
   chembl_id_lookup \tab Look up ChEMBL Id entity type \cr
   compound_record \tab Occurence of a given compound in a spcecific document \cr
   compound_structural_alert \tab Indicates certain anomaly in compound structure \cr
   document \tab Document/Dataset from which Assays have been extracted \cr
   document_similarity \tab Provides documents similar to a given one \cr
   document_term \tab Provides keywords extracted from a document using the TextRank algorithm \cr
   drug \tab Approved drugs information, including (but not limited to) applicants, patent numbers and research codes. This endpoint aggregates data on the parent, please use the parent chembl id found in other endpoints \cr
   drug_indication \tab Joins drugs with diseases providing references to relevant sources \cr
   drug_warning \tab Safety information for drugs withdrawn from one or more regions of the world and drugs that carry a warning for severe or life threatening adverse effects \cr
   go_slim \tab GO slim ontology \cr
   image \tab Graphical (svg) representation of Molecule \cr
   mechanism \tab Mechanism of action information for approved drugs \cr
   metabolism \tab Metabolic pathways with references \cr
   molecule \tab Molecule information, including properties, structural representations and synonyms \cr
   molecule_form \tab Relationships between molecule parents and salts \cr
   organism \tab Simple organism classification \cr
   protein_classification \tab Protein family classification of TargetComponents \cr
   similarity \tab Molecule similarity search \cr
   source \tab Document/Dataset source \cr
   status \tab API status with ChEMBL DB version number and API software version number \cr
   substructure \tab Molecule substructure search \cr
   target \tab Targets (protein and non-protein) defined in Assay \cr
   target_component \tab Target sequence information (A Target may have 1 or more sequences) \cr
   target_relation \tab Describes relations between targets \cr
   tissue \tab Tissue classification \cr
   xref_source \tab Cross references to other resources for compounds \cr
}
\tabular{cc}{
   Filter Type \tab Description \cr
   exact (iexact) \tab Exact match with query (case insensitive equivalent) \cr
   contains (icontains) \tab Wild card search with query (case insensitive equivalent) \cr
   startswith (istartswith) \tab Starts with query (case insensitive equivalent) \cr
   endswith (iendswith) \tab Ends with query (case insensitive equivalent) \cr
   regex (iregex) \tab Regular expression query (case insensitive equivalent) \cr
   gt (gte) \tab Greater than (or equal) \cr
   lt (lte) \tab Less than (or equal) \cr
   range \tab Within a range of values \cr
   in \tab Appears within list of query values \cr
   isnull \tab Field is null \cr
   search \tab Special type of filter allowing a full text search based on elastic search queries \cr
   only \tab Select specific properties from the original endpoint and returns only the desired properties on each record \cr
}
}
