% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctrp_treatmentIDs}
\alias{ctrp_treatmentIDs}
\title{ctrp_treatmentIDs is a preprocessed version of the CTRP treatment metadata.}
\format{
A data table with 2 columns and 545 rows.
\describe{
\item{CTRP.treatmentid}{\code{char} The treatment id of the all the compounds used in the experiment.}
\item{CTRP.broad_cpd_id}{\code{char} The broad compound id of the all the compounds used in the experiment.}
}
}
\source{
https://ctd2-data.nci.nih.gov/Public/Broad/CTRPv2.0_2015_ctd2_ExpandedDataset/CTRPv2.0_2015_ctd2_ExpandedDataset.zip
}
\usage{
data(ctrp_treatmentIDs)
}
\description{
Data was obtained from "https://ctd2-data.nci.nih.gov/Public/Broad/CTRPv2.0_2015_ctd2_ExpandedDataset/CTRPv2.0_2015_ctd2_ExpandedDataset.zip"
and preprocessed to only get two columns: CTRP.treatmentid and CTRP.broad_cpd_id.
}
\examples{
data(ctrp_treatmentIDs)
head(ctrp_treatmentIDs)
}
\keyword{datasets}
