% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodoAPI.R
\name{depositZenodo}
\alias{depositZenodo}
\alias{deposit_zenodo}
\title{Upload and optionally publish data to Zenodo}
\usage{
depositZenodo(
  file_path,
  metadata = zenodoMetadata(),
  publish = FALSE,
  ...,
  url = "https://zenodo.org/api/deposit/depositions",
  access_token = Sys.getenv("ZENODO_TOKEN")
)
}
\arguments{
\item{file_path}{\code{character(1)} Path to the file to upload to Zenodo.}

\item{metadata}{\code{list} Named list of metadata required to publish publish
a Zenodo entry, as returned by the \code{zenodoMetadata} function. See
\code{?zenodoMetadata} for documentation of the correct format.}

\item{publish}{\code{logical(1)} Should the Zenodo entry be published, assuming
the upload succeeds? Default is \code{FALSE}.}

\item{...}{Fall through parameters to the \code{httr::POST} and \code{httr::PUT}
method calls inside this function. Pass \code{httr::verbose()} to get detailed
information about your API requests, useful for debugging.}

\item{url}{\code{character(1)} URL string for the API. Do not change unless you
know what you are doing.}

\item{access_token}{\code{character(1)} Zenodo access token. By deafult tries to
read this from the ZENODO_TOKEN environmet variable via \code{Sys.getenv}.
See details for more information.}
}
\value{
If all HTTP requests succeed, returns the details of the Zenodo entry
as a list invisiblty. If any of the requests fail, it returns the failed
request object for ease of debugging.
}
\description{
Make the PUT and POST request necessary to upload and optionally publish
a file to Zenodo.
}
\details{
\subsection{Caveats}{

This function is an early prototype and is not very polished. The current
version will create a new entry everytime the function is called, even
if some of the steps succeeded. For now, you will need to manually delete
these entries from you Zenodo account. However, this will be corrected in
future updates.
}

\subsection{\code{access_token}}{

API documentation is available at https://developers.zenodo.org/.

To use this function you need a Zenodo access token for your account. A link
is available in the API documentation to create one. It is assumed you have
added the access token to your operating system environment in the
ZENODO_TOKEN variable.

If you have not, you can specify an access token. For this function to
work the token needs to provide both upload and publish permissions on your
Zenodo account.
}

\subsection{Trouble Shooting}{
\itemize{
\item \emph{405 Errors}: Ensure that you have verified your Zenodo account email.
\item \emph{403 Errors}: Ensure you have checked both the \code{publish:actions} and
\code{publish:write} boxes for your Zenodo access token.
}
}
}
\seealso{
\link[httr:POST]{httr::POST}, \link[httr:upload_file]{httr::upload_file}
}
