% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem_rest.R
\name{.build_pubchem_rest_query}
\alias{.build_pubchem_rest_query}
\title{Build a query for the PubChem REST API}
\usage{
.build_pubchem_rest_query(
  id,
  domain = "compound",
  namespace = "name",
  operation = "cids",
  output = "JSON",
  url = "https://pubchem.ncbi.nlm.nih.gov/rest/pug",
  raw = FALSE,
  query_only = FALSE,
  ...
)
}
\arguments{
\item{id}{The identifier(s) for the query. If namespace is 'name', id must be a single value.}

\item{domain}{The domain of the query. Options are 'compound', 'substance', 'assay', 'cell', 'gene', 'protein'.}

\item{namespace}{The namespace of the query. Options depend on the chosen domain.}

\item{operation}{The operation to perform. Options depend on the chosen domain and namespace.}

\item{output}{The desired output format. Options are 'JSON', 'XML', 'SDF', 'TXT', 'CSV'.}

\item{url}{The base URL for the PubChem REST API.}

\item{raw}{Logical indicating whether to return the raw response or parse it.}

\item{query_only}{Logical indicating whether to return the query URL only.}

\item{...}{Additional arguments to be passed to the query.}
}
\value{
The query URL or the parsed response, depending on the arguments.
}
\description{
This function builds a query for the PubChem REST API based on the provided parameters.
}
\keyword{internal}
