% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem_view.R
\name{.build_pubchem_view_query}
\alias{.build_pubchem_view_query}
\title{Build a PubChem REST query URL}
\usage{
.build_pubchem_view_query(
  id,
  annotation = "data",
  record = "compound",
  page = NULL,
  version = NULL,
  heading = NULL,
  source = NULL,
  output = "JSON",
  ...
)
}
\arguments{
\item{id}{The identifier for the query.}

\item{annotation}{The type of annotation to retrieve. Options include "data", "index", "annotations", "categories", "neighbors", "literature", "structure", "image", "qr", or "linkout".}

\item{record}{The type of record to retrieve. Options include "compound", "substance", "assay", "cell", "gene", or "protein".}

\item{page}{The page number to retrieve. Retrieves the nth page of the output data when the data is paginated. By default, retrieve all pages of the output data.}

\item{version}{The version of the record to retrieve. For substance, this is an integer. For assay, this is a string in the format "n.m".}

\item{heading}{The name of the annotation heading or subheading to retrieve.}

\item{source}{The name of the annotation source to retrieve from a specified source.}

\item{output}{The desired output format. Options are "JSON", "XML", "SDF", "TXT", "CSV".}
}
\value{
The query URL
}
\description{
pubchem\if{html}{\out{<DOT>}}ncbi\if{html}{\out{<DOT>}}nlm\if{html}{\out{<DOT>}}nih\if{html}{\out{<DOT>}}gov/rest/pug_view/\if{html}{\out{<ANNOTATION>}}/\if{html}{\out{<RECORD>}}/\if{html}{\out{<ID>}}/\if{html}{\out{<OUTPUT>}}<?OPTIONS>
}
\details{
Optional Parameters:
\itemize{
\item annotation: data, index, annotations, categories, neighbors, literature, structure, image, qr, linkout
\item record: compound, substance, assay, cell, gene, protein
\item filter (Parameter, Value, Description):
\itemize{
\item Page, n \if{html}{\out{<integer>}}, The page number to retrieve. Retrieves the nth page of the output data when the data is paginated.
This option is useful when downloading the annotations under a specific heading for all compounds
By default, retrieve all pages of the output data.
\item Version, n \if{html}{\out{<int>}} for substance | n.m <int.int> for assay, data for a particular version of a substance or assay record
\item Heading, (Sub)heading name, The name of the annotation heading or subheading to retrieve
\item Source, source name, The name of the annotation source to retrieve from a specified source
}
}
}
\keyword{internal}
