% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.buildURL}
\alias{.buildURL}
\title{Helper function to connect}
\usage{
.buildURL(...)
}
\arguments{
\item{...}{Arguments passed through to \code{paste}. The \code{sep} arguement is
already set to '/'.}
}
\value{
A \code{character} vector with the URL.
}
\description{
Helper function to connect
}
