% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChem.R
\name{.checkThrottlingStatus}
\alias{.checkThrottlingStatus}
\title{Checks to see if the PubChem query is exceeding the throttling limit}
\usage{
.checkThrottlingStatus(response)
}
\arguments{
\item{response}{\code{httr::response}}
}
\description{
Checks to see if the PubChem query is exceeding the throttling limit
}
