% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem_helpers.R
\name{.checkThrottlingStatus2}
\alias{.checkThrottlingStatus2}
\title{names are: request_count, request_time and service
each has status and percent
main throttlers for user are request_count and request_time
main statuses are:
Green - less than 50\% of the permitted request limit has been used
Yellow - between 50\% and 75\% of the request limit has been used
Red - more than 75\% of the request limit has been reached
Black - the limit has been exceeded and requests are being blocked}
\usage{
.checkThrottlingStatus2(response, printMessage)
}
\description{
names are: request_count, request_time and service
each has status and percent
main throttlers for user are request_count and request_time
main statuses are:
Green - less than 50\% of the permitted request limit has been used
Yellow - between 50\% and 75\% of the request limit has been used
Red - more than 75\% of the request limit has been reached
Black - the limit has been exceeded and requests are being blocked
}
\keyword{internal}
