% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellosaurusAPI.R
\name{.constructCellosaurusURL}
\alias{.constructCellosaurusURL}
\title{Construct Cellosaurus API URL}
\usage{
.constructCellosaurusURL(query, from, extResource, to, numResults)
}
\description{
Construct Cellosaurus API URL
}
\keyword{internal}
