% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.error}
\alias{.error}
\title{Custom error function with formatted message}
\usage{
.error(...)
}
\description{
This function is used to raise an error with a formatted message.
The message is printed in magenta and bold format.
}
\keyword{internal}
