% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.formatMessage}
\alias{.formatMessage}
\title{Format a message with optional collapsing}
\usage{
.formatMessage(..., collapse = ", ")
}
\description{
This function takes multiple arguments and concatenates them into a single string.
The arguments are separated by a collapse character, which is a comma by default.
The resulting string is then wrapped to a new line if it exceeds a certain width.
}
\keyword{internal}
