% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.funContext}
\alias{.funContext}
\title{Generate a function context string}
\usage{
.funContext(funName)
}
\description{
This function takes the name of a function and returns a string that
represents the function context.
The string is formatted as \link{packageName functionName}.
}
\keyword{internal}
