% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem_view.R
\name{.get_all_heading_types}
\alias{.get_all_heading_types}
\title{https://pubchem.ncbi.nlm.nih.gov/rest/pug/annotations/headings/JSON will return a list of all available headings}
\usage{
.get_all_heading_types()
}
\description{
https://pubchem.ncbi.nlm.nih.gov/rest/pug/annotations/headings/JSON will return a list of all available headings
}
