% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.groupListByName}
\alias{.groupListByName}
\title{Collect repeated list item names into a single list item}
\usage{
.groupListByName(dataList)
}
\arguments{
\item{dataList}{A \code{list} with repeated names.}
}
\value{
A \code{list} with unique names, collecting repeats into sublists.
}
\description{
Collect repeated list item names into a single list item
}
