% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.message}
\alias{.message}
\title{Custom message function for verbose output}
\usage{
.message(...)
}
\description{
This function is used to print messages when the verbose option is enabled.
It checks if the package-specific verbose option is set or if the global verbose option is set.
If either of these options is TRUE, the message is printed in blue and bold format.
}
\keyword{internal}
