% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{.parse_cellosaurus_lines}
\alias{.parse_cellosaurus_lines}
\title{Parses the lines of a cellosaurus response}
\usage{
.parse_cellosaurus_lines(resp)
}
\arguments{
\item{resp}{The response object containing the cellosaurus data}
}
\value{
A list of parsed lines from the cellosaurus data
}
\description{
This function takes a response object and parses the lines of the response
to extract specific sections of the cellosaurus data.
}
