% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem_rest.R
\name{.parse_pubchem_rest_responses}
\alias{.parse_pubchem_rest_responses}
\title{Map compound names to PubChem CIDs}
\usage{
.parse_pubchem_rest_responses(responses)
}
\description{
Map compound names to PubChem CIDs
}
