% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_httr2_utils.R
\name{.parse_resp_json}
\alias{.parse_resp_json}
\title{Parses the JSON response from an HTTP request.}
\usage{
.parse_resp_json(resp)
}
\arguments{
\item{resp}{The response object from the HTTP request.}
}
\value{
The parsed JSON response.
}
\description{
Parses the JSON response from an HTTP request.
}
