% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChem.R
\name{.parseATCresponse}
\alias{.parseATCresponse}
\title{Function that parses the results of the PubChem PUG-VIEW API for the ATC Code header}
\usage{
.parseATCresponse(result)
}
\description{
Function that parses the results of the PubChem PUG-VIEW API for the ATC Code header
}
