% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChem.R
\name{.parseCASresponse}
\alias{.parseCASresponse}
\title{Function that parses the results of the PubChem PUG-VIEW API for the CAS header}
\usage{
.parseCASresponse(result)
}
\arguments{
\item{result}{The result object obtained from the PubChem PUG-VIEW API.}
}
\value{
The CAS header information extracted from the result object.
}
\description{
This function takes the result object obtained from the PubChem PUG-VIEW API and extracts the CAS header information.
}
