% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellosaurusAPI.R
\name{.parseCellosaurusTSVResponse}
\alias{.parseCellosaurusTSVResponse}
\title{Parse Cellosaurus TSV Response}
\usage{
.parseCellosaurusTSVResponse(response)
}
\arguments{
\item{response}{The response from the Cellosaurus API.}
}
\value{
A data.table containing the parsed information from the response.
}
\description{
This function parses the response from the Cellosaurus API and converts it into a data.table.
}
