% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellosaurusAPI.R
\name{.parseCellosaurusTXTResponse}
\alias{.parseCellosaurusTXTResponse}
\title{Parses the response from the Cellosaurus API.}
\usage{
.parseCellosaurusTXTResponse(response)
}
\arguments{
\item{response}{The response from the Cellosaurus API.}
}
\value{
A data.table object containing the parsed data.
}
\description{
This function takes the response from the Cellosaurus API and parses it into a structured format.
It removes unnecessary header lines, splits the response into individual records, and extracts the relevant fields.
The resulting data is returned as a data.table object.
}
