% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChem.R
\name{.parseCHEMBLresponse}
\alias{.parseCHEMBLresponse}
\title{Function that parses the results of the PubChem PUG-VIEW API for the CHEMBL ID header}
\usage{
.parseCHEMBLresponse(result)
}
\arguments{
\item{result}{The result object obtained from the PubChem PUG-VIEW API.}
}
\value{
The CHEMBL ID header extracted from the result object.
}
\description{
This function takes the result object obtained from the PubChem PUG-VIEW API
and extracts the CHEMBL ID header from it.
It removes the "Compound::" prefix from the header and returns the cleaned
result.
}
