% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_httr2_utils.R
\name{.perform_request}
\alias{.perform_request}
\title{Performs an HTTP request.}
\usage{
.perform_request(request)
}
\arguments{
\item{request}{The HTTP request to perform.}
}
\value{
The response of the HTTP request.
}
\description{
Performs an HTTP request.
}
