% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{.split_cellosaurus_lines}
\alias{.split_cellosaurus_lines}
\title{Splits cellosaurus lines into a named list}
\usage{
.split_cellosaurus_lines(lines)
}
\arguments{
\item{lines}{A vector of cellosaurus lines}
}
\value{
A named list where each element corresponds to a unique identifier and contains the associated lines
}
\description{
This function takes a vector of cellosaurus lines and splits them into a named list.
The lines are split based on the delimiter "   " (three spaces).
}
\examples{
lines <- c("ID1   Line 1", "ID1   Line 2", "ID2   Line 1", "ID2   Line 2")
.split_cellosaurus_lines(lines)
# Output:
# $ID1
# [1] "Line 1" "Line 2"
#
# $ID2
# [1] "Line 1" "Line 2"

}
