% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{.strSplitFinite}
\alias{.strSplitFinite}
\title{Split a string into a finite number of capture groups}
\usage{
.strSplitFinite(x, split, n, fixed)
}
\description{
Split a string into a finite number of capture groups
}
