% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{.strSplitInfinite}
\alias{.strSplitInfinite}
\title{Split a string into an finite number of capture groups}
\usage{
.strSplitInfinite(x, split, fixed)
}
\description{
Split a string into an finite number of capture groups
}
