% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGencode.R
\name{.valid_gencode_files}
\alias{.valid_gencode_files}
\title{Retrieve valid file arguments for getGencodeFile}
\usage{
.valid_gencode_files(type = c("GTF", "FASTA", "metadata"))
}
\description{
Retrieve valid file arguments for getGencodeFile
}
\keyword{internal}
