% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.warning}
\alias{.warning}
\title{Custom warning function with formatted message}
\usage{
.warning(...)
}
\description{
This function is used to raise a warning with a formatted message.
The message is printed in cyan and bold format.
}
\keyword{internal}
