% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downloadAndExtract}
\alias{downloadAndExtract}
\title{Download a compressed file from a remote URL and extract it.}
\usage{
downloadAndExtract(url, extract_fun = unzip, ...)
}
\arguments{
\item{url}{\code{character(1)} URL of the compressed file to download.}

\item{extract_fun}{\code{character(1)} or \code{function} to unzip the downloaded
file with. Default is \code{utils::unzip}.}

\item{...}{Fall through arguments to \code{extract_fun}. See documentation of
the specified \code{extract_fun} for more details.}
}
\value{
\code{character} vector of unzipped file paths when \code{extract_fun} is
\code{unzip} (default), otherwise the return value of the specified
\code{extract_fun}.
}
\description{
Download a compressed file from a remote URL and extract it.
}
\seealso{
\link[utils:unzip]{utils::unzip}, \link[utils:untar]{utils::untar}, \link[R.utils:gunzip]{R.utils::gunzip}, \link[R.utils:bunzip2]{R.utils::bunzip2}
}
