% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFDA.R
\name{downloadFDAOrangeBook}
\alias{downloadFDAOrangeBook}
\title{Download and extract the FDA OrangeBook files from the FDA website}
\usage{
downloadFDAOrangeBook(
  url = "https://www.fda.gov/media/76860/download",
  output_dir = tempdir()
)
}
\arguments{
\item{url}{\code{character(1)} URL to download the OrangeBook files from.}

\item{output_dir}{``}
}
\value{
\code{character} vector of paths to the FDA OrangeBook files.
}
\description{
Download and extract the FDA OrangeBook files from the FDA website
}
