% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsquash.R
\name{fsquash}
\alias{fsquash}
\title{Recursively flatten a \code{data.frame} while correctly handling nested lists
of \code{data.frame}s within \code{list} columns.}
\usage{
fsquash(x)
}
\arguments{
\item{x}{\code{data.frame} Table to squash, recursively flattening list columns
and nested \code{data.frame}s.}
}
\value{
\code{data.frame} The data in \code{x}, but with only atomic column types.
Values are repeated and mappings retained if the nested colums contain
multiple entrier per row.
}
\description{
Recursively flatten a \code{data.frame} while correctly handling nested lists
of \code{data.frame}s within \code{list} columns.
}
