% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationGx-data.R
\docType{data}
\name{gCSI_sampleMetadata}
\alias{gCSI_sampleMetadata}
\title{gCSI Sample Metadata}
\format{
A \code{data.table} with 1350 rows and 2 columns:
\describe{
\item{Characteristics.cell.line.}{gCSI cell line name.}
\item{Comment.ENA_SAMPLE.}{ENA sample accession.}
}
}
\source{
gCSI source metadata; BHKLab-curated PSet on ORCESTRA,
\url{https://www.orcestra.ca/pset/10.5281/zenodo.4737437}. See
\code{system.file("extdata", "LICENSES.md", package = "AnnotationGx")} for
bundled data provenance notes.
}
\usage{
data(gCSI_sampleMetadata)
}
\description{
A column-reduced example sample metadata table derived from Genentech Cell
Line Screening Initiative (gCSI) metadata.
}
\examples{
data(gCSI_sampleMetadata)
head(gCSI_sampleMetadata)
}
\keyword{datasets}
