% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationGx-data.R
\docType{data}
\name{GDSC_sampleMetadata}
\alias{GDSC_sampleMetadata}
\title{GDSC Sample Metadata}
\format{
A \code{data.table} with 1001 rows and 2 columns:
\describe{
\item{GDSC.Sample_Name}{GDSC cell line name.}
\item{GDSC.COSMIC_ID}{COSMIC identifier.}
}
}
\source{
GDSC, \url{https://www.cancerrxgene.org/}; BHKLab-curated PSet on
ORCESTRA, \url{https://www.orcestra.ca/pset/10.5281/zenodo.3905481}. See
\code{system.file("extdata", "LICENSES.md", package = "AnnotationGx")} for
bundled data provenance notes.
}
\usage{
data(GDSC_sampleMetadata)
}
\description{
A column-reduced example sample metadata table derived from the Genomics of
Drug Sensitivity in Cancer (GDSC) data.
}
\examples{
data(GDSC_sampleMetadata)
head(GDSC_sampleMetadata)
}
\keyword{datasets}
