% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationGx-data.R
\docType{data}
\name{GDSC_treatmentMetadata}
\alias{GDSC_treatmentMetadata}
\title{GDSC Treatment Metadata}
\format{
A \code{data.table} with 621 rows and 3 columns:
\describe{
\item{GDSC.treatmentid}{GDSC treatment name.}
\item{GDSC.synonyms}{GDSC-provided treatment synonyms.}
\item{GDSC.drug_id}{GDSC drug identifier.}
}
}
\source{
GDSC, \url{https://www.cancerrxgene.org/}; BHKLab-curated PSet on
ORCESTRA, \url{https://www.orcestra.ca/pset/10.5281/zenodo.3905481}. See
\code{system.file("extdata", "LICENSES.md", package = "AnnotationGx")} for
bundled data provenance notes.
}
\usage{
data(GDSC_treatmentMetadata)
}
\description{
A column-reduced example treatment metadata table derived from GDSC.
}
\examples{
data(GDSC_treatmentMetadata)
head(GDSC_treatmentMetadata)
}
\keyword{datasets}
