% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-webScraping.R
\name{get_remote_table}
\alias{get_remote_table}
\title{Get all HTML tables from a remote URL.}
\usage{
get_remote_table(url)
}
\arguments{
\item{url}{\code{character(1)} URL to scrape all HTML tables from.}
}
\description{
This function takes a URL as input and scrapes all HTML tables from that URL.
}
