% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem_view.R
\name{getAnnotationHeadings}
\alias{getAnnotationHeadings}
\title{Get annotation headings based on type and heading criteria.}
\usage{
getAnnotationHeadings(type = "all", heading = NULL)
}
\arguments{
\item{type}{The type of annotation headings to retrieve. Options include "Compound", "Gene", "Taxonomy", "Element", "Assay", "Protein", "Cell", "Pathway", or "all" (default).}

\item{heading}{The specific heading to filter the results by. Defaults to NULL, which retrieves all headings.}
}
\value{
A data frame containing the annotation headings that match the specified criteria.
}
\description{
Get annotation headings based on type and heading criteria.
}
\examples{
getAnnotationHeadings()
getAnnotationHeadings(type = "Compound")
getAnnotationHeadings(heading = "ChEMBL*") # get all headings that contain "ChEMBL"

}
