% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBrainArrayCDFs.R
\name{getBrainArrayTable}
\alias{getBrainArrayTable}
\title{Retrieve a table of all available CDF files from BrainArray}
\usage{
getBrainArrayTable(
 
    url = "http://brainarray.mbni.med.umich.edu/Brainarray/Database/CustomCDF/25.0.0/ensg.asp"
)
}
\arguments{
\item{`character(1)`}{URL to a page listing CDF files on BrainArray. Defaults
to the ENSG version for Ensembl IDs.}
}
\description{
Retrieve a table of all available CDF files from BrainArray
}
\examples{
brain_array <- getBrainArrayTable()

}
