% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryCellosaurus.R
\name{getCellApi}
\alias{getCellApi}
\title{Query Cellosaurus}
\usage{
getCellApi(cl_names)
}
\arguments{
\item{cl_names}{is a list of the cell line names}
}
\value{
A list of responses
}
\description{
This function takes a list of cell line names and gets responses from the Cellosaurus API
}
\details{
Function to get responses from Cellosaurus API
}
