% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellosaurusAPI.R
\name{getCellosaurusAccesions}
\alias{getCellosaurusAccesions}
\title{getCellosaurusAccesions Function}
\usage{
getCellosaurusAccesions(samples, from = "idsy", ..., threads = 1)
}
\arguments{
\item{samples}{A character vector of sample IDs.}

\item{from}{The source of the sample IDs. Default is "idsy".}

\item{...}{Additional arguments to be passed to the searchCellosaurusAPI function.}

\item{threads}{The number of threads to use for parallel processing. Default is 1.}
}
\value{
A data.table object containing the Cellosaurus accessions for the given samples.
}
\description{
This function retrieves Cellosaurus accessions for a given set of samples.
}
\examples{
getCellosaurusAccesions(samples = c("sample1", "sample2"), from = "idsy", threads = 2)

}
