% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellosaurusAPI.R
\name{getCellosaurusAPI}
\alias{getCellosaurusAPI}
\title{Query Cellosaurus}
\usage{
getCellosaurusAPI(
  cl_names,
  fields = c("id", "ac", "sy", "misspelling", "din", "ca", "sx", "ag", "sampling-site",
    "metastatic-site"),
  GETfxn = c("search/cell-line?", "cell-line/"),
  querydomain = "ac:"
)
}
\arguments{
\item{cl_names}{is a list of the cell line names}

\item{fields}{is a list of desired fields to obtain for each cell line in the API query, i.e if only trying to get synonynms and primary accesssion then fields=c("sy", "ac"). see https://api.cellosaurus.org/static/fields_help.html for all fields.}
}
\value{
A list of responses
}
\description{
This function takes a list of cell line names and interested fields and gets responses from the Cellosaurus API
}
\details{
Function to get responses from Cellosaurus API
}
