% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellosaurus.R
\name{getCelloxml}
\alias{getCelloxml}
\title{Access and read xml}
\usage{
getCelloxml(
  url = "https://ftp.expasy.org/databases/cellosaurus/cellosaurus.xml"
)
}
\arguments{
\item{url}{is cellosaurus link to xml}
}
\description{
Access and read xml
}
