% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChEMBL.R
\name{getChemblAllMechanisms}
\alias{getChemblAllMechanisms}
\title{Retrieve a \code{data.table} of all molecule mechanisms of action from ChEMBL}
\usage{
getChemblAllMechanisms(
  url = "https://www.ebi.ac.uk",
  query = "/chembl/api/data/mechanism.json?limit=1000"
)
}
\arguments{
\item{url}{\code{character(1)} Base URL of ChEMBL API.}

\item{query}{\code{character(1)} ChEMBL API query.}
}
\value{
\code{data.table} A table containing all mechanism of action entries
availble in ChEMBL.
}
\description{
Retrieve a \code{data.table} of all molecule mechanisms of action from ChEMBL
}
