% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chembl.R
\name{getChemblFilterTypes}
\alias{getChemblFilterTypes}
\title{Get the Chembl filter types}
\usage{
getChemblFilterTypes()
}
\value{
A list of Chembl filter types.
}
\description{
This function retrieves the Chembl filter types.
}
\examples{
getChemblFilterTypes()

}
