% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chembl.R
\name{getChemblMechanism}
\alias{getChemblMechanism}
\title{Get ChEMBL Mechanism}
\usage{
getChemblMechanism(
  chembl.ID,
  resources = "mechanism",
  field = "molecule_chembl_id",
  filter_type = "in",
  returnURL = FALSE,
  raw = FALSE
)
}
\arguments{
\item{chembl.ID}{The ChEMBL ID of the molecule.}

\item{resources}{The ChEMBL resource to query (default: "mechanism").}

\item{field}{The field to filter on (default: "molecule_chembl_id").}

\item{filter_type}{The filter type to use (default: "in").}

\item{returnURL}{Logical indicating whether to return the constructed URL (default: FALSE).}

\item{raw}{Logical indicating whether to return the raw response JSON (default: FALSE).}
}
\value{
A data.table containing the retrieved mechanism information.
}
\description{
This function retrieves information about the mechanism of action for a given ChEMBL ID.
}
\examples{
getChemblMechanism("CHEMBL1413")
getChemblMechanism("CHEMBL1413", resources = "mechanism", field = "molecule_chembl_id",
                   filter_type = "in", returnURL = FALSE, raw = FALSE)


}
