% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chembl.R
\name{getChemblResourceFields}
\alias{getChemblResourceFields}
\title{Get the fields of a Chembl resource}
\usage{
getChemblResourceFields(resource)
}
\arguments{
\item{resource}{The Chembl resource.}
}
\value{
A character vector containing the names of the fields.
}
\description{
This function retrieves the fields of a Chembl resource.
}
\examples{
getChemblResourceFields("molecule")

}
