% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chembl.R
\name{getChemblResources}
\alias{getChemblResources}
\title{getChemblResources function}
\usage{
getChemblResources()
}
\value{
A list of Chembl resources.
}
\description{
This function retrieves the Chembl resources.
}
\examples{
getChemblResources()

}
