% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellosaurus.R
\name{getCompleteCellosaurusObject}
\alias{getCompleteCellosaurusObject}
\title{Fetches the complete Cellosaurus data}
\usage{
getCompleteCellosaurusObject()
}
\value{
A data table containing the complete Cellosaurus data
}
\description{
This function fetches the complete Cellosaurus data and converts it
into a data table using the Cellosaurus package.
}
