% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrugBank.R
\name{getDrugTargets}
\alias{getDrugTargets}
\title{Extract all drug targets from the \code{target} node of the DrugBank .XML file}
\usage{
getDrugTargets(filePath, progressbars = TRUE, ..., BPPARAM = bpparam())
}
\arguments{
\item{filePath}{\code{character(1)} Path to the Drug Bank .XML file.}

\item{progressbars}{\code{logical(1)} Should progress bars be shown for
parallelized parts of this functions execution.}

\item{...}{Force subequent parameters to be named. Not used.}

\item{BPPARM}{Configurate parallelization backend. See
\code{\link[BiocParallel:register]{BiocParallel::bpparam}} for details.}
}
\value{
A \code{data.table} containing all drug target data from the .XML file
as a long format table. List columns are collapsed into string where items
are delimited by '|' and names are captured as 'name=value' for each
item.
}
\description{
Extract all drug targets from the \code{target} node of the DrugBank .XML file
}
