% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getFailed}
\alias{getFailed}
\title{Get the failed attribute of an object}
\usage{
getFailed(x)
}
\arguments{
\item{x}{An object}
}
\value{
The "failed" attribute of the object
}
\description{
This function retrieves the "failed" attribute of an object.
}
