% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getFailedIDs}
\alias{getFailedIDs}
\title{Get the failed IDs from an object}
\usage{
getFailedIDs(x)
}
\arguments{
\item{x}{An object}
}
\value{
A vector of failed IDs
}
\description{
This function retrieves the "failed" IDs from an object by accessing the "query" element of the "failed" attribute.
}
