% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getFailureMessages}
\alias{getFailureMessages}
\title{Get failure messages from an object}
\usage{
getFailureMessages(x)
}
\arguments{
\item{x}{An object with a 'failed' attribute}
}
\value{
A data frame containing the query and corresponding failure messages
}
\description{
This function retrieves failure messages from an object that has a 'failed' attribute.
}
