% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFDA.R
\name{getFDAOrangeBookProducts}
\alias{getFDAOrangeBookProducts}
\title{Download and load the FDA OrangeBook products file}
\usage{
getFDAOrangeBookProducts(file_path)
}
\arguments{
\item{file_path}{\code{character} Optional path to the FDA OrangeBook products
file. If excluded the file is downloaded from the FDA website.}
}
\value{
\code{data.table} FDA OrangeBook products table.
}
\description{
Download and load the FDA OrangeBook products file
}
