% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGencode.R
\name{getGencodeAvailableFiles}
\alias{getGencodeAvailableFiles}
\title{Retrieve a list of files and their descriptions available for a Gencode
release and reference genome version.}
\usage{
getGencodeAvailableFiles(
  version = "latest",
  chr = c("GRCh38", "GRCh37"),
  dir = tempdir(),
  url = "https://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human"
)
}
\arguments{
\item{version}{\code{character(1)} Gencode version to download for.
Defaults to "latest". See \code{getGencodeFTPTable()} for options.
Versions prior to 10 are not currently supported.}

\item{chr}{\code{character(1)} Name of reference chromosome to fetch files for.
Options are "GRCh38" (default) and "GRCh37".}

\item{dir}{\code{character(1)} Path to download the file to. Defaults to
\code{tempdir()}. When this value is \code{tempdir()} the downloaded file is
automatically deleted when the function exits.}

\item{url}{\code{character(1)} Address of Gencode FTP web page. Default is page
for Gencode Human files.}
}
\value{
\code{data.table} With columns type, file, and description.
}
\description{
Retrieve a list of files and their descriptions available for a Gencode
release and reference genome version.
}
