% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGencode.R
\name{getGencodeFile}
\alias{getGencodeFile}
\title{Download files from the Gencode FTP site and load them as the appropriate
Bioconductor classes.}
\usage{
getGencodeFile(
  file = "comprehensive_chr",
  type = "GTF",
  version = "latest",
  dir = tempdir(),
  url = "https://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human"
)
}
\arguments{
\item{file}{\code{character(1)} Character vector of files to download from the
Gencode FTP site. Defaults to "comprehensive_chr". See details for options.}

\item{type}{\code{character(1)} One of "GTF", "FASTA" or "metadata". Defaults to
"GTF".}

\item{version}{\code{character(1)} Gencode version to download for.
Defaults to "latest". See \code{getGencodeFTPTable()} for options.}

\item{dir}{\code{character(1)} Path to download the file to. Defaults to
\code{tempdir()}. When this value is \code{tempdir()} the downloaded file is
automatically deleted when the function exits.}

\item{url}{\code{character(1)} Address of Gencode FTP web page. Default is page
for Gencode Human files.}
}
\value{
\code{GenomcRanges} object when type="GTF", \code{DNAStringSet} when
type="FASTA", or \code{data.table}/\code{character} (as appropriate) when
type="metadata".
}
\description{
File descriptions are available at https://www.gencodegenes.org/human/.
}
\details{
\subsection{Options for \code{file}}{

When type="GTF" or "GFF":
\itemize{
\item comprehensive_all
\item comprehensive_chr
\item comprehensive_primary
\item basic_all
\item basic_chr
\item long_non_coding
\item poly_a
\item consensus_pseudogenes
\item predicted_trna
}

When type="FASTA":
\itemize{
\item transcript
\item protein_coding
\item long_non_coding
\item genome_all
\item genome_primary
}

When type="metadata":
\itemize{
\item remarks
\item entrez
\item exon_evidence
\item gene_source
\item pdb
\item poly_a
\item pubmed_id
\item ref_seq
\item selenocyteine
\item swiss_prot
\item transcript_source
\item transcipt_evidence
\item trembl
}
}
}
