% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGencode.R
\name{getGencodeFilesTable}
\alias{getGencodeFilesTable}
\title{Retrieve a table listing all files available from Gencode for a specific
database release.}
\usage{
getGencodeFilesTable(
  version = "latest",
  url = "https://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human"
)
}
\arguments{
\item{version}{\code{character(1)} Version number of Gencode annotations to
retrieve file list for. Defaults to "latest", which is the most recent
release. Use \code{getGencodeFTPTable()} to view available releases.}
}
\value{
\code{data.table} Table listing all files available for the selected
\code{version}, along with URLs to download each file from.
}
\description{
Retrieve a table listing all files available from Gencode for a specific
database release.
}
\seealso{
Other Gencode: 
\code{\link{getGencodeAvailableFiles}()},
\code{\link{getGencodeFTPTable}()},
\code{\link{getGencodeFile}()},
\code{\link{getGencodeGRangesAnnotated}()},
\code{\link{infer_gencode_type}()}
}
\concept{Gencode}
