% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGencode.R
\name{getGencodeFTPTable}
\alias{getGencodeFTPTable}
\title{Retrieve the index table for the Gencode FTP web page}
\usage{
getGencodeFTPTable(
  url = "https://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human",
  recursive = FALSE
)
}
\arguments{
\item{url}{\code{character(1)} Address of Gencode FTP web page. Default is page
for Gencode Human files.}

\item{recursive}{\code{logical(1)} Should the file tables be recursively retrieved
for directories? Default is \code{FALSE}, setting to \code{TRUE} will be be slow due
to the large number of directories to scrape.}
}
\value{
\code{data.table} Table of files and directories available on the
Gencode FTP web page.
}
\description{
Retrieve the index table for the Gencode FTP web page
}
