% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGencode.R
\name{getGencodeGRangesAnnotated}
\alias{getGencodeGRangesAnnotated}
\title{Retrieve a GRanges object for a specific Gencode release with added metadata}
\usage{
getGencodeGRangesAnnotated(annotation = "SwissProt", ...)
}
\arguments{
\item{annotation}{\code{character(1)} Name of the annotation to retrieve. Currently
only supports, and defaults to, "SwissProt". More will be implemented in
the future.}

\item{...}{\code{pairlist()} Fall through arguments to \code{getGencodeFile} when
retreiving the annotaton files. See \code{?getGencodeFile} for details.}
}
\value{
\code{GRanges} File of Gencode genome annotations for the file retreived
using \code{getGencodeFile(...)} with annotations added from the selected
Gencode metadatafile.
}
\description{
Retrieve a GRanges object for a specific Gencode release with added metadata
}
