% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGuideToPharm.R
\name{getGuideToPharm}
\alias{getGuideToPharm}
\title{Get data from the Guide to PHARMACOLOGY Database Web Services}
\usage{
getGuideToPharm(
  ids = NA,
  service = "ligands",
  id_type = "name",
  ...,
  url = "https://www.guidetopharmacology.org/services"
)
}
\arguments{
\item{ids}{\code{character()} or \code{integer()} Identifiers to query the web
service with. If excluded, the entire record for the specified service
is returned.}

\item{service}{\code{character(1)} Which Guide to PHARMACOLOGY web service
to query. Defaults to 'ligands'. Other options are 'targets', 'interactions',
'diseases' and 'references'.}

\item{id_type}{\code{character(1)} What type of identifiers are in \code{ids}? Defaults
to 'name', for drug name. Other options are 'accession', which accepts
PubChem CIDs.}

\item{...}{Force subsequent parameters to be named. Not used.}

\item{url}{\code{character(1)} The URL of the Guide to PHARMACOLOGY API. Do
not change this unless you are a developer and know what you are doing.}
}
\description{
Get data from the Guide to PHARMACOLOGY Database Web Services
}
\details{
The API reference documentation can be found here:
https://www.guidetopharmacology.org/webServices.jsp

There is also a Python interface available for querying this API. See:
https://github.com/samirelanduk/pygtop
}
