% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellosaurus.R
\name{getInfoFromCellID}
\alias{getInfoFromCellID}
\title{Filter parent node cell-line and parse child nodes for required annotations}
\usage{
getInfoFromCellID(cell_nm, main_xml)
}
\arguments{
\item{cell_nm}{is cell name}

\item{main_xml}{is read xml object from getCelloxml}
}
\description{
Filter parent node cell-line and parse child nodes for required annotations
}
