% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellosaurus.R
\name{getInfoFromCelllineInput}
\alias{getInfoFromCelllineInput}
\title{Filter parent node cell-line and parse child nodes for required annotations}
\usage{
getInfoFromCelllineInput(cell_ip, main_xml, input_type, remove = FALSE)
}
\arguments{
\item{cell_ip}{is either cell name or cvcl id.}

\item{main_xml}{is read xml object from getCelloxml}

\item{input_type}{is either cell line name(s) denoted by \code{name} or Cellosaurus ID (CVCL ID) denoted by \code{cvclid}.}
}
\value{
A \code{list} of xml child nodes
}
\description{
Filter parent node cell-line and parse child nodes for required annotations
}
