% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncotree.R
\name{getOncotreeVersions}
\alias{getOncotreeVersions}
\title{Get available Oncotree versions}
\usage{
getOncotreeVersions()
}
\value{
A \code{data.table} containing available Oncotree versions.
}
\description{
This function retrieves the available versions of Oncotree.
}
\examples{
# Requires internet connection to Oncotree API
if (interactive()) {
  getOncotreeVersions()
}

}
