% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrugBank.R
\name{getPathways}
\alias{getPathways}
\title{Extract the \code{pathway} item from a node in the DrugBank .xml file}
\usage{
getPathways(nodeL)
}
\arguments{
\item{nodeL}{A \code{list} produced by calling \code{xml2::as_list} on a single node
in the DrugBank database .xml file.}
}
\value{
A \code{data.table} containing the pathway data for \code{nodeL}
}
\description{
Extract the \code{pathway} item from a node in the DrugBank .xml file
}
