% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChem.R
\name{getPubChemAnnotation}
\alias{getPubChemAnnotation}
\title{getPubChemAnnotation}
\usage{
getPubChemAnnotation(
  compound,
  annotationType = "ChEMBL ID",
  url = "https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound",
  timeout_s = 29,
  retries = 3,
  quiet = TRUE,
  throttleMessage = FALSE,
  query_only = FALSE
)
}
\arguments{
\item{compound}{\code{character(1)} A valid CID to use for the query.}

\item{url}{\code{character(1)} The URL to perform API queries on. default = 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound'}

\item{timeout_s}{\code{numeric(1)} The number of seconds to wait before timing out. Default is 29.}

\item{retries}{\code{numeric(1)} The number of times to retry a failed query. Default is 3.}

\item{quiet}{\code{logical(1)} Should the function be quiet? Default is TRUE.}

\item{throttleMessage}{\code{logical(1)} Should a message be printed when the query is throttled? Default is FALSE.}

\item{query_only}{\code{logical(1)} Should this function early return only the encoded query?}

\item{header}{\code{character(1)} A valid header name for the PUG VIEW annotations}

\item{output}{\code{character(1)} The output format. Defaults to 'JSON'.}
}
\value{
A \code{data.table} of resulting annotations.
}
\description{
queries the PubChem PUG-VIEW API to get a single annotation using a CID for a header
}
