% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChem.R
\name{getPubChemAnnotations}
\alias{getPubChemAnnotations}
\title{Get annotations for the entire PubChem database}
\usage{
getPubChemAnnotations(
  header = "Available",
  type = "Compound",
  ...,
  output = "JSON",
  raw = FALSE,
  url = "https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/annotations/heading",
  parseFUN = identity
)
}
\arguments{
\item{header}{\code{character(1)} A valid header name for the PUG VIEW annotations
API. Default is 'Available', which will return a list of available
headers as a \code{data.frame}.}

\item{type}{\code{character(1)} The header type. Default is 'Compound'. Make
sure ot change this if your header of interest isn't type compouns.}

\item{...}{Force subsequent parameters to be named. Not used.}
}
\description{
Get annotations for the entire PubChem database
}
\section{API Documentation}{
For detailed documentation of the annotations API see:
https://pubchemdocs.ncbi.nlm.nih.gov/pug-view$_Toc495044630
}

