% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChem.R
\name{getPubChemCompound}
\alias{getPubChemCompound}
\title{getPubChemCompound}
\usage{
getPubChemCompound(
  ids,
  from = "cid",
  to = "property",
  ...,
  properties = "Title",
  batch = TRUE,
  raw = FALSE,
  proxy = FALSE,
  options = NA,
  query_only = FALSE
)
}
\arguments{
\item{ids}{A \code{character} or \code{numeric} vector of valid PubChem identifiers
to use for the query. Which identifier is being used must be specified in
the \code{from} parameter.}

\item{from}{A \code{character(1)} vector with the desired namespace to query.
Default is 'cid'. Try using 'sid' if some of your CIDs fail to map.}

\item{to}{A \code{character(1)} vector with the desired return type. Defaults
to 'record', which returns all available data from the specified IDs.}

\item{...}{Fallthrough arguments to \code{BiocParallel::bmapply}.}

\item{properties}{A \code{character} vector of properties to return. Only used
when \code{to='property'}. Common properties of interest are: 'Title' (name),
'IUPACName', 'CanonicalSMILES', 'IsomericSMILES', 'InChIKey'. The default
setting will return 'Title'.
See details for more information.}

\item{batch}{\code{logical(1)} Should the query be run in batches (i.e., multiple
ids per GET request to the API). Default is \code{TRUE}. Should be set to
\code{FALSE} when retrying failed queries. Batch queries are not supported when
\code{from} is one of 'name', 'xref', 'smiles', 'inchi' or 'sdf'. In these
cases, batch will automatically be set to \code{FALSE} with a warning.}

\item{raw}{\code{logical(1)} Should the raw query results be early returned. This
can be useful for diagnosing issues with failing queries.}

\item{proxy}{\code{logical(1)} Route API queries through random proxy servers?
this can increase query length, but it useful if you have been blacklisted.}
}
\value{
A \code{data.table} containing results of the query, or a list if \code{raw}
is set to \code{TRUE}. Failed queries are available as an attribute of the
returned object, see \code{attributes(object)}.
}
\description{
Make queries to the PubChem Compound domain.
}
\details{
\subsection{\code{properties}}{

For a full list of availabe properties see:
https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest$_Toc494865556
}
}
