% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChem.R
\name{getPubChemCompoundFromCID}
\alias{getPubChemCompoundFromCID}
\title{getPubChemFromCID}
\usage{
getPubChemCompoundFromCID(
  ids,
  to = "property",
  ...,
  properties = c("Title", "IUPACName", "CanonicalSMILES", "IsomericSMILES", "InChIKey"),
  batch = TRUE,
  raw = FALSE
)
}
\arguments{
\item{ids}{A \code{character} or \code{numeric} vector of valid PubChem CIDs to use
for the query.}

\item{to}{A \code{character(1)} vector with the desired return type. Defaults
to 'record', which returns all available data from the specified IDs.}

\item{...}{Force subsequent parameters to be named. Not used.}

\item{properties}{A \code{character} vector of properties to return. Only used
when \code{to='property'}. Common properties of interest are: 'Title' (name),
'IUPACName', 'CanonicalSMILES', 'IsomericSMILES', 'InChIKey'. The default
setting with return all of these.
See details for more information.}
}
\value{
A \code{data.frame} or \code{list} containing results of the query.
}
\description{
getPubChemFromCID
}
\details{
\subsection{\code{properties}}{

For a full list of availabe properties see:
https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest$_Toc494865556
}
}
