% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChem.R
\name{getPubChemFromNSC}
\alias{getPubChemFromNSC}
\title{getPubChemFromNSC}
\usage{
getPubChemFromNSC(ids, to = "cids", ..., raw = FALSE)
}
\arguments{
\item{ids}{A \code{character} or \code{numeric} vector of valid NSC ids to use for the
query.}

\item{to}{A \code{character(1)} vector with the desired return type. Currently
only 'cids' and 'sids' are implemented, but other options are available
via the PubChem API.}

\item{...}{Force subsequent parameters to be named. Not used.}

\item{raw}{A \code{logical(1)} vector specifying whether to early return the raw
query results. Use this if specifying an unimplemented return to the \code{to}
paramater.}
}
\value{
A \code{data.frame} where the first column is the specified NSC ids and
the second column is the results specified in \code{to}.
}
\description{
Return a data.frame mapping from ids to the information specified in \code{to}.
}
