% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChem-helpers.R
\name{getPubChemStatus}
\alias{getPubChemStatus}
\title{simple checkThrottling Status by a default response}
\usage{
getPubChemStatus(sleep = FALSE)
}
\value{
\code{logical} whether the query is throttled
}
\description{
simple checkThrottling Status by a default response
}
