% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChem.R
\name{getPubChemSubstance}
\alias{getPubChemSubstance}
\title{getPubChemSubstance}
\usage{
getPubChemSubstance(
  ids,
  from = "cid",
  to = "sids",
  ...,
  batch = TRUE,
  raw = FALSE,
  proxy = FALSE
)
}
\arguments{
\item{ids}{A \code{character} or \code{numeric} vector of valid PubChem identifiers
to use for the query. Which identifier is being used must be specified in
the \code{from} parameter.}

\item{from}{A \code{character(1)} vector with the desired namespace to query.
Default is 'cid'. Try using 'sid' if some of your CIDs fail to map.}

\item{to}{A \code{character(1)} vector with the desired return type. Defaults
to 'record', which returns all available data from the specified IDs.}

\item{batch}{\code{logical(1)} Should the query be run in batches (i.e., multiple
ids per GET request to the API). Default is \code{TRUE}. Should be set to
\code{FALSE} when retrying failed queries. Batch queries are not supported when
\code{from} is one of 'name', 'xref', 'smiles', 'inchi' or 'sdf'. In these
cases, batch will automatically be set to \code{FALSE} with a warning.}

\item{raw}{\code{logical(1)} Should the raw query results be early returned. This
can be useful for diagnosing issues with failing queries.}

\item{proxy}{}
}
\value{
A \code{data.frame} or \code{list} containing results of the query.
}
\description{
Make queries to the PubChem Compound domain.
}
