% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChem.R
\name{getRequestPubChem}
\alias{getRequestPubChem}
\title{Constructs and executes a GET request to the PubChem PUG REST API}
\usage{
getRequestPubChem(
  id,
  domain = "compound",
  namespace = "cid",
  operation = NA,
  output = "JSON",
  ...,
  url = "https://pubchem.ncbi.nlm.nih.gov/rest/pug",
  operation_options = NA,
  raw = FALSE,
  query_only = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{id}{The identifier for REST API query. This should be a valid compound
identifier for the specified \code{domain} and \code{namespace}. For example, if
\code{domain='compound'} and \code{namespace='cid'} then \code{id} should be one or
more compound cids to retrieve information about. If \code{id} is a \code{list} or
\code{vector}, the ids are parsed to a string with ids separated by
commas. This parameter forms the '\if{html}{\out{<identifiers>}}' part of the Input
portion of the URL path: '\if{html}{\out{<domain>}}/\if{html}{\out{<namespace>}}/\if{html}{\out{<identifiers>}}'.}

\item{domain}{Which records in PubChem are the subject of the query? Default
is 'compound'. Usually, this will be one of 'substance', 'compound', or
'assay'. For more advanced options see the PubChem PUG REST API
documentation. This forms the '\if{html}{\out{<domain>}}' part of the Input portion of the
URL path.}

\item{namespace}{What kind of identifiers are in \code{id}? These options
are specific to the selected domain. For the 'compound' domain, common
options are 'cid', 'name', 'smiles', 'inchi', or 'formula'. For the
'substance' domain common options are 'sid' or 'name'. For the
'assay' domain common options are 'aid' and 'target/<assay_target>' where
'<assay_target>' is one of 'gi', 'proteinname', 'geneid', 'genesymbol' or
'accession'. For all options please see the PubChem PUG REST API
documentation. This parameter makes up the '\if{html}{\out{<namespace>}}' part of the
Input portion of the URL path.}

\item{operation}{What kind of data to return for the specified input
portion of the URL path? If excluded this defaults to return the entire
record for that specified input. Options are domain specific and details
are available in the PubChem PUG REST API documentation. Options valid for
'compound', 'substance' and 'assay' domains are 'record', 'aids', 'sids',
'aids', 'description' and 'classification'. Common options for the
'compound' domain are 'assaysummary', 'conformers' and
'property/\if{html}{\out{<properties>}}' where '\if{html}{\out{<properties>}}' is a comma separated list
property tags; valid tags can be found at
https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest$_Toc494865565.
Common options for the 'substance' domain are 'assaysummary'.
Common options for the 'assay' domain include 'concise', 'summary' and
'targets/\if{html}{\out{<target type>}}' where '\if{html}{\out{<target type>}}' is one of 'ProteinGI',
'ProteinName', 'GeneID' or 'GeneSymbol'.}

\item{output}{What format should the data be returned in? Default is 'JSON'.
other common options include 'XML', 'CSV' and 'TXT'. For a complete list of
output format options, plase see the PubChem PUG REST API documentation.}

\item{...}{Fall through arguments to \code{\link[httr:GET]{httr::GET}}.}

\item{url}{The URL of the PubChem REST API. Probably don't change this.}

\item{operation_options}{Further optional arguments for the selected operation.
this is specific to the selected operation. This is appended as a string
after '?' at the end of the query. See
https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest$_Toc494865565 for details.}

\item{query_only}{\code{logical(1)} Should this function early return only
the encoded query?}
}
\value{
A \code{httr::response} object with the results of the GET request.
}
\description{
This function builds a query URL for the PubChem PUG REST API based on the
function parameters then executes that query, returning a \code{httr::request}
object.

Most – if not all – of the information PubChem PUG service needs to
produce its results is encoded into the URL. The general form of the URL
has three parts – input, operation, and output – after the common prefix,
followed by operation options as URL arguments (after the ‘?’):

"https://pubchem.ncbi.nlm.nih.gov/rest/pug/\if{html}{\out{<input specification>}}/\if{html}{\out{<operation specification>}}/[\if{html}{\out{<output specification>}}][?<operation_options>]"
\subsection{Input}{

The input portion of the URL tells the service which records to use as the
subject of the query. This is further subdivided into two or more locations
in the URL “path” as follows:

"\if{html}{\out{<input specification>}} = \if{html}{\out{<domain>}}/\if{html}{\out{<namespace>}}/\if{html}{\out{<identifiers>}}"

"\if{html}{\out{<domain>}} = substance | compound | assay | \if{html}{\out{<other inputs>}}"

"compound domain \if{html}{\out{<namespace>}} = cid | name | smiles | inchi | sdf | inchikey | formula | \if{html}{\out{<structure search>}} | \if{html}{\out{<xref>}} | listkey | \if{html}{\out{<fast search>}}"

"substance domain \if{html}{\out{<namespace>}} = sid | sourceid/\if{html}{\out{<source id>}} | sourceall/\if{html}{\out{<source name>}} | name | \if{html}{\out{<xref>}} | listkey"

"assay domain \if{html}{\out{<namespace>}} = aid | listkey | type/\if{html}{\out{<assay type>}} | sourceall/\if{html}{\out{<source name>}} | target/\if{html}{\out{<assay target>}} | activity/\if{html}{\out{<activity column name>}}"

Complete documentation for valid input specifications can be found at:
https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest$_Toc494865556
}

\subsection{Operation}{

The operation part of the URL tells the service what to do with the input
records – such as to retrieve whole record data blobs or specific
properties of a compound, etc. The construction of this part of the
“path” will depend on what the operation is. Currently, if no operation is
specified at all, the default is to retrieve the entire record. What
operations are available are, of course, dependent on the input domain –
that is, certain operations are applicable only to compounds and not
assays.

"compound domain \if{html}{\out{<operation specification>}} = record | \if{html}{\out{<compound property>}} | synonyms | sids | cids | aids | assaysummary | classification | \if{html}{\out{<xrefs>}} | description | conformers"

"substance domain \if{html}{\out{<operation specification>}} = record | synonyms | sids | cids | aids | assaysummary | classification | \if{html}{\out{<xrefs>}} | description"

"assay domain \if{html}{\out{<operation specification>}} = record | concise | aids | sids | cids | description | targets/\if{html}{\out{<target type>}} | \if{html}{\out{<doseresponse>}} | summary | classification"

Complete documentation for valid operations can be found at:
https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest$_Toc494865557
}

\subsection{Output}{

The final portion of the URL tells the service what output format is desired.
Note that this is formally optional, as output format can also be specified
in the HTTP Accept field of the request header – see below for more detail.

"\if{html}{\out{<output specification>}} = XML | ASNT | ASNB | JSON | JSONP  | SDF | CSV | PNG | TXT"

Complete documentation for valid output formats can be found at:
https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest$_Toc494865558
}
}
\details{
See the PUG_VIEW website for the complete
API documentation. A subset of this documentation is included below for
convenience.
}
\references{
Kim S, Thiessen PA, Cheng T, Yu B, Bolton EE. An update on PUG-REST: RESTful interface for programmatic access to PubChem. Nucleic Acids Res. 2018 July 2; 46(W1):W563-570. doi:10.1093/nar/gky294.

Kim S, Thiessen PA, Bolton EE, Bryant SH. PUG-SOAP and PUG-REST: web services for programmatic access to chemical information in PubChem. Nucleic Acids Res. 2015 Jul 1; 43(W1):W605-W611. doi: 10.1093/nar/gkv396.

Kim S, Thiessen PA, Bolton EE. Programmatic Retrieval of Small Molecule Information from PubChem Using PUG-REST. In Kutchukian PS, ed. Chemical Biology Informatics and Modeling. Methods in Pharmacology and Toxicology. New York, NY: Humana Press, 2018, pp. 1-24. doi:10.1007/7653_2018_30.
}
\seealso{
\link[httr:GET]{httr::GET}, \link[httr:RETRY]{httr::RETRY}, \link{queryRequestPubChem}
}
