% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrugBank.R
\name{getTargets}
\alias{getTargets}
\title{Extract the \code{target} item from a node in the DrugBank .xml file}
\usage{
getTargets(nodeL)
}
\arguments{
\item{nodeL}{A \code{list} produced by calling \code{xml2::as_list} on a single node
in the DrugBank database .xml file.}
}
\value{
A \code{data.table} containing the target data for \code{nodeL}
}
\description{
Extract the \code{target} item from a node in the DrugBank .xml file
}
