% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-webScraping.R
\name{grep_directory_names}
\alias{grep_directory_names}
\title{Extracts directory names from a given vector of paths.}
\usage{
grep_directory_names(x)
}
\arguments{
\item{x}{A vector of paths.}
}
\value{
A vector of directory names.
}
\description{
Extracts directory names from a given vector of paths.
}
\examples{
grep_directory_names(c("/home/user/dir1/file1.txt", "/home/user/dir2/file2.txt"))

}
