% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-webScraping.R
\name{grep_file_names}
\alias{grep_file_names}
\title{Find file names matching specified extensions in a vector of file paths.}
\usage{
grep_file_names(x, extensions = "[[:alnum:]]{2,5}$")
}
\arguments{
\item{x}{\code{character()} vector of file or directory paths}

\item{extensions}{\code{character(1)} regex of file extensions to match. Defaults
to files which end with 2 to 5 alphanumeric characters preceded by a dot.}
}
\description{
Find file names matching specified extensions in a vector of file paths.
}
