% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniChem.R
\name{identifierToInchikey}
\alias{identifierToInchikey}
\title{Map from a database identifier to an InchiKey via the UniChem API}
\usage{
identifierToInchikey(
  chemical_id,
  target_names,
  ...,
  base_url = "https://www.ebi.ac.uk/unichem/rest/structure/",
  inchikey = TRUE
)
}
\arguments{
\item{chemical_id}{A \code{character} vector which is the src_compound_id}

\item{src_id}{A \code{character} vector which is the id for the database/source}
}
\value{
A \code{character} vector with the inchikey and inchi structure
}
\description{
This function takes in chemical id and source id and returns the
inchi and inchikey structure
}
