% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_ignore.R
\name{ignore_unused_imports}
\alias{ignore_unused_imports}
\title{ignore_unused_imports function}
\usage{
ignore_unused_imports()
}
\value{
None
}
\description{
This function is used to ignore unused imports.
It calls the crayon::green function and the httr2::request function.
}
\keyword{internal}
