% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniChem.R
\name{inchiKeyToIdentifier}
\alias{inchiKeyToIdentifier}
\title{Map from InchiKey to a database specific identifier via the UniChem API}
\usage{
inchiKeyToIdentifier(
  inchi_key,
  ...,
  base_url = "https://www.ebi.ac.uk/unichem/rest/inchikey/"
)
}
\arguments{
\item{inchikey}{A \code{character} vector which is unique to each compound.}
}
\value{
A data.frame with source id and source compound id.
}
\description{
This function takes in inchikey as a parameter and allows to map
to any identifier using the UniChem API
}
