% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniChem.R
\name{inchiToDatabaseID}
\alias{inchiToDatabaseID}
\title{Map from an InchiKey to a database ID using the UniChem API}
\usage{
inchiToDatabaseID(
  inchi,
  target_names,
  ...,
  type = c("key", "structure", ),
  base_url = "https://www.ebi.ac.uk/unichem/rest"
)
}
\arguments{
\item{inchi}{A \code{character(1)} vector containing the InchiKey if \code{type}="key"}

\item{target_names}{A \code{character()} vector specifying the name of the target
database id to return. If missing, returns all}
}
\value{
A \code{data.frame} mapping from the inchi or inchikey to the
}
\description{
This function takes in inchikey and names of databases and returns the
appropriate chemical source ids
}
