% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGencode.R
\name{infer_gencode_type}
\alias{infer_gencode_type}
\title{Guess the type of a Gencode file from its name}
\usage{
infer_gencode_type(file)
}
\arguments{
\item{file}{\code{character(1)} File name.}
}
\value{
\code{character(1)} One of "GTF", "GFF3", "FASTA" or "metadata" based on
the file extensions of \code{file}.
}
\description{
Guess the type of a Gencode file from its name
}
