% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrugBank.R
\name{listColToDT}
\alias{listColToDT}
\title{Convert a list column in a \code{data.table} to a \code{data.table}}
\usage{
listColToDT(col)
}
\arguments{
\item{col}{A \code{list} column in a \code{data.tabel}}
}
\value{
A \code{data.table} produced by calling \code{as.data.table} on the list column
}
\description{
Convert a list column in a \code{data.table} to a \code{data.table}
}
