% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniChem.R
\name{mapBetweenSources}
\alias{mapBetweenSources}
\title{Map from one database ID to another database ID via the UniChem API}
\usage{
mapBetweenSources(
  chemical_id,
  src_name,
  target_name,
  ...,
  end_point = "src_compound_id",
  base_url = "https://www.ebi.ac.uk/unichem/rest"
)
}
\arguments{
\item{chemical_id}{A \code{character} vector which is the compound identifier for
the specified database/source}

\item{src_name}{A \code{character} vector which is the short name for the database/
source for which we know the chemical_id}

\item{target_name}{A \code{character} vector which is the short name for the database/
source for which we require the compounds end_point}

\item{...}{Force subsequent parameters to be named. Not used.}

\item{end_point}{A \code{character} vector which is the src_compound_id, the
id of that compound for that particular database/source}

\item{base_url}{A \code{character} vector specifying the base_url to use when constructing
API queries. This is for developer use.}
}
\value{
A \code{character} vector which is the src_compound_id for the target_name
}
\description{
This function allows to map between external databases identifiers using the UniChem API.
The UniChem database gathers its data from 48 different sources/databases. All sources are listed with their
src_id and short name.

|src#| src_name        |
| 1  | chembl          |
| 2  | drugbank        |
| 3  | pdb             |
| 4  | gtopdb          |
| 5  | pubchem_dotf    |
| 6  | kegg_ligand     |
| 7  | chebi           |
| 8  | nih_ncc         |
| 9  | zinc            |
| 10 | emolecules      |
| 11 | ibm             |
| 12 | atlas           |
| 14 | fdasrs          |
| 15 | surechembl      |
| 17 | pharmakb        |
| 18 | hmdb            |
| 20 | selleck         |
| 21 | pubchem_tpharma |
| 22 | pubchem         |
| 23 | mcule           |
| 24 | nmrshiftdb2     |
| 25 | lincs           |
| 26 | actor           |
| 27 | recon           |
| 28 | molport         |
| 29 | nikkaji         |
| 31 | bindingDB       |
| 32 | comptox         |
| 33 | lipidmaps       |
| 34 | drugcentral     |
| 35 | carotenoiddb    |
| 36 | metabolights    |
| 37 | brenda          |
| 38 | rhea            |
| 39 | chemicalbook    |
| 40 | dailymed_old    |
| 41 | swiss_lipids    |
| 45 | dailymed        |
| 46 | clinicaltrials  |
| 47 | rxnorm          |
| 48 | MedChemExpress  |
}
