% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{mapCell2Accession}
\alias{mapCell2Accession}
\title{Map Cell Line IDs to Accession Numbers}
\usage{
mapCell2Accession(
  ids,
  numResults = 10000,
  from = "idsy",
  sort = "ac",
  keep_duplicates = FALSE,
  fuzzy = FALSE,
  query_only = FALSE,
  raw = FALSE,
  parsed = TRUE,
  include_query = TRUE,
  ...
)
}
\arguments{
\item{ids}{A character vector of cell line IDs.}

\item{numResults}{The maximum number of results to return for each query. Default is 1000.}

\item{from}{The type of input IDs. Possible values are "idsy" (default), "ac", "id", "sy", and "misspelling".}

\item{sort}{The sorting order of the results. Possible values are "ac" (default), "id", "sy", and "misspelling".}

\item{keep_duplicates}{Logical indicating whether to keep duplicate results. Default is FALSE.}

\item{fuzzy}{Logical indicating whether to perform a fuzzy search. Default is FALSE.}

\item{query_only}{Logical indicating whether to return only the query URLs. Default is FALSE.}

\item{raw}{Logical indicating whether to return the raw HTTP responses. Default is FALSE.}

\item{parsed}{Logical indicating whether to parse the response text. Default is TRUE.}

\item{include_query}{Logical indicating whether to include the \verb{query*} columns
(e.g. \code{query}, \code{query:ac}) in the returned result. Default is TRUE.}

\item{...}{Currently unused. Reserved for future extensions.}
}
\value{
A data.table containing the mapped cell line IDs and accession numbers.
When \code{parsed = FALSE}, the returned table also includes Cellosaurus metadata
columns that have been renamed to user-friendly titles (for example, \code{sy}
becomes \code{synonyms}).
}
\description{
This function maps cell line IDs to accession numbers using the Cellosaurus database.
}
\examples{
mapCell2Accession(ids = c("A549", "MCF7"))

}
