% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellosaurusAPI.R
\name{mapCellosaursAccessionsToFields}
\alias{mapCellosaursAccessionsToFields}
\title{Maps Cellosaurus accessions to specified fields}
\usage{
mapCellosaursAccessionsToFields(accessions, fields, threads = 1)
}
\arguments{
\item{accessions}{A vector of Cellosaurus accessions.}

\item{fields}{A vector of fields to retrieve for each accession.}

\item{threads}{The number of threads to use for parallel processing.}
}
\value{
A data.table containing the retrieved information for each accession.
}
\description{
This function takes a vector of Cellosaurus accessions and a vector of fields,
and retrieves the corresponding information from the Cellosaurus API. It uses
multiple cores for parallel processing, with the number of threads specified
by the 'threads' parameter.
}
