% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem_rest.R
\name{mapCID2Properties}
\alias{mapCID2Properties}
\title{Map PubChem Compound IDs to Properties}
\usage{
mapCID2Properties(ids, properties, ...)
}
\arguments{
\item{ids}{A vector of PubChem Compound IDs.}

\item{properties}{A vector of property names to retrieve for each compound.}

\item{...}{Additional arguments to be passed to the \code{getPubchemCompound} function.}
}
\value{
A data frame containing the mapped properties for each compound.
}
\description{
This function maps PubChem Compound IDs to specified properties using the PubChem REST API.
See \code{getPubchemProperties} for a list of available properties.
}
\examples{
mapCID2Properties(ids = c(123, 456), properties = c("MolecularWeight", "CanonicalSMILES"))

}
