% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem_rest.R
\name{mapCompound2CID}
\alias{mapCompound2CID}
\title{Map compound names to PubChem CIDs}
\usage{
mapCompound2CID(
  names,
  raw = FALSE,
  query_only = FALSE,
  output = "JSON",
  first = FALSE,
  ...
)
}
\arguments{
\item{names}{A character vector of compound names.}

\item{raw}{Logical indicating whether to return the raw response from the API (default is FALSE).}

\item{query_only}{Logical indicating whether to only perform the query without retrieving the data (default is FALSE).}

\item{output}{The format of the output, either 'JSON' or 'XML' (default is 'JSON').}

\item{...}{Additional arguments to be passed to the getPubchemCompound function.}
}
\value{
A character vector of PubChem CIDs.
}
\description{
This function maps compound names to PubChem CIDs using the PubChem REST API.
}
\examples{
mapCompound2CID(c("aspirin", "caffeine"))

}
