% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query_CTD2.R
\name{mapCompound2CTD}
\alias{mapCompound2CTD}
\title{Map Compound to CTD}
\usage{
mapCompound2CTD(
  drug,
  base_url = "https://ctd2-dashboard.nci.nih.gov/dashboard/get",
  endpoint = "compound",
  raw = FALSE,
  query_only = FALSE
)
}
\arguments{
\item{drug}{The name of the drug compound to be mapped.}

\item{base_url}{The base URL of the CTD API. Default is "https://ctd2-dashboard.nci.nih.gov/dashboard/get".}

\item{endpoint}{The API endpoint for the compound mapping. Default is "compound".}

\item{raw}{Logical indicating whether to return the raw response from the API. Default is FALSE.}

\item{query_only}{Logical indicating whether to only return the API request URL without making the actual request. Default is FALSE.}
}
\value{
A data table containing the mapped information for the drug compound.
If the API request fails, a data table with the drug compound name will be returned.
If \code{raw} is set to TRUE, the raw response from the API will be returned.
}
\description{
This function maps a drug compound to the Cancer Target Discovery (CTD) database.
It retrieves information about the compound from the CTD database and returns the results as a data table.
}
\examples{
mapCompound2CTD("Bax channel blocker")

}
