% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchNested-methods.R
\name{matchNested}
\alias{matchNested}
\alias{matchNested,character,list-method}
\alias{matchNested,numeric,list-method}
\alias{matchNested,character,data.table-method}
\alias{matchNested,numeric,data.table-method}
\alias{matchNested,character,data.frame-method}
\title{Match inside nested elements}
\usage{
matchNested(x, table, ..., keep_duplicates = FALSE)

\S4method{matchNested}{character,list}(x, table, keep_duplicates)

\S4method{matchNested}{numeric,list}(x, table, keep_duplicates)

\S4method{matchNested}{character,data.table}(x, table, keep_duplicates)

\S4method{matchNested}{numeric,data.table}(x, table, keep_duplicates)

\S4method{matchNested}{character,data.frame}(x, table, keep_duplicates)

\S4method{matchNested}{character,data.frame}(x, table, keep_duplicates)
}
\arguments{
\item{x}{The values to be matched.}

\item{table}{The values to be matched against.
Applies across rows for \code{DataFrame} method.}

\item{...}{Additional arguments to be passed to the method.}

\item{keep_duplicates}{A logical value indicating whether to keep duplicates.}
}
\value{
\code{integer}.
A positional vector corresponding to values defined in \code{table} the same
size as \code{x}.
}
\description{
Match inside nested elements
}
\details{
Intentionally only performs exact matching. Refer to \code{filterNested} function
for partial matching support with regular expressions.
}
\examples{
showMethods("matchNested")
}
