% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChEMBL.R
\name{moleculeQuery}
\alias{moleculeQuery}
\title{Query the ChEMBL molecule API end-point}
\usage{
moleculeQuery(field, filter_type, value)
}
\description{
\tabular{cc}{
   Field \tab Subfields \cr
   atc_classifications \tab  \cr
   availability_type \tab  \cr
   biotherapeutic \tab  \cr
   black_box_warning \tab  \cr
   chebi_par_id \tab  \cr
   chirality \tab  \cr
   cross_references \tab  \cr
   dosed_ingredient \tab  \cr
   first_approval \tab  \cr
   first_in_class \tab  \cr
   helm_notation \tab  \cr
   indication_class \tab  \cr
   inorganic_flag \tab  \cr
   max_phase \tab  \cr
   molecule_chembl_id \tab  \cr
   molecule_hierarchy \tab molecule_chembl_id, parent_chembl_id, \cr
   molecule_properties \tab alogp, aromatic_rings, cx_logd, cx_logp, cx_most_apka, cx_most_bpka, full_molformula, full_mwt, hba, hba_lipinski, hbd, hbd_lipinski, heavy_atoms, molecular_species, mw_freebase, mw_monoisotopic, num_lipinski_ro5_violations, num_ro5_violations, psa, qed_weighted, ro3_pass, rtb \cr
   molecule_structures \tab canonical_smiles, molfile, standard_inchi, standard_inchi_key \cr
   molecule_synonyms \tab  \cr
   molecule_type \tab  \cr
   natural_product \tab  \cr
   oral \tab  \cr
   parenteral \tab  \cr
   polymer_flag \tab  \cr
   pref_name \tab  \cr
   pro_drug \tab  \cr
   structure_type \tab  \cr
   therapeutic_flag \tab  \cr
   topical \tab  \cr
   usan_stem \tab  \cr
   usan_stem_definition \tab  \cr
   usan_substem \tab  \cr
   usan_year \tab  \cr
   withdrawn_class \tab  \cr
   withdrawn_country \tab  \cr
   withdrawn_flag \tab  \cr
   withdrawn_reason \tab  \cr
   withdrawn_year \tab  \cr
}
}
