% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrugBank.R
\name{parseDrugBankXML}
\alias{parseDrugBankXML}
\title{Read in DrugBank database from an XML file as a list of data.tables}
\usage{
parseDrugBankXML(
  xmlPath,
  table = c("pathways", "targets"),
  outPath = tempdir(),
  ...,
  nodeFun
)
}
\arguments{
\item{xmlPath}{a}

\item{outPath}{\code{character}}

\item{...}{a}
}
\description{
\strong{WARNING}: Due to the size of the database dump it is likely this
function will crash your computer if you have less than 32 GB of RAM!
}
