% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProxyManager-class.R
\name{ProxyManager}
\alias{ProxyManager}
\title{Singleton ProxyManager Class to Manage a List of Proxy URLs}
\description{
Singleton ProxyManager Class to Manage a List of Proxy URLs

Singleton ProxyManager Class to Manage a List of Proxy URLs
}
\section{Fields}{
name \code{character(1)} A string name for the ProxyManager. Defaults to
'ProxyManger'
}

\section{Super class}{
\code{\link[R6P:Singleton]{R6P::Singleton}} -> \code{ProxyManager}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProxyManager$new()}}
\item \href{#method-get_proxies}{\code{ProxyManager$get_proxies()}}
\item \href{#method-update_failed_proxies}{\code{ProxyManager$update_failed_proxies()}}
\item \href{#method-get_failed_proxies}{\code{ProxyManager$get_failed_proxies()}}
\item \href{#method-clone}{\code{ProxyManager$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProxyManager$new(
 
    proxy_source_url = "https://api.proxyscrape.com/v2/?request=displayproxies&protocol=http&timeout=10000&country=all&ssl=all&anonymity=all",
  name = "ProxyManager",
  ...,
  proxy_parser = parse_text_response_to_datatable
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_proxies"></a>}}
\if{latex}{\out{\hypertarget{method-get_proxies}{}}}
\subsection{Method \code{get_proxies()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProxyManager$get_proxies()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_failed_proxies"></a>}}
\if{latex}{\out{\hypertarget{method-update_failed_proxies}{}}}
\subsection{Method \code{update_failed_proxies()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProxyManager$update_failed_proxies(new_data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_failed_proxies"></a>}}
\if{latex}{\out{\hypertarget{method-get_failed_proxies}{}}}
\subsection{Method \code{get_failed_proxies()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProxyManager$get_failed_proxies()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProxyManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
