% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryCellosaurus.R
\name{queryCellosaurus}
\alias{queryCellosaurus}
\title{Access and parse Cellosaurus xml for developers}
\usage{
queryCellosaurus(
  url = "https://ftp.expasy.org/databases/cellosaurus/cellosaurus.xml",
  cellline_input,
  namespace = "name",
  verbose = TRUE
)
}
\arguments{
\item{url}{is cellosaurus link to xml. This should be a valid link to Cellosaurus xml. Default is \verb{https://ftp.expasy.org/databases/cellosaurus/cellosaurus.xml}.}

\item{cellline_input}{is any cell line identifier. Cell line name(s) or `Cellosaurus ID (CVCL ID)`` can be provided as the input.}

\item{namespace}{is either cell line name (default) or Cellosaurus ID (CVCL ID).}

\item{verbose}{is TRUE by default}
}
\value{
An \code{XML} document of Cellosaurus
}
\description{
This function reads Cellosaurus XML and parses parent node for each cell line, returning an XML document.
This enables parsing any child node of interest.
}
\details{
This function reads Cellosaurus XML and parses parent node for each cell line, returning an XML document.
}
\references{
Bairoch A.The Cellosaurus, a cell line knowledge resource.J. Biomol. Tech. 29:25-38(2018) DOI: 10.7171/jbt.18-2902-002; PMID: 29805321; PMCID: PMC5945021
}
