% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chembl.R
\name{queryChemblAPI}
\alias{queryChemblAPI}
\title{Query the ChEMBL API}
\usage{
queryChemblAPI(resource, field, filter_type, value, format = "json")
}
\arguments{
\item{resource}{The resource to query in the ChEMBL API.}

\item{field}{The field to filter on in the ChEMBL API.}

\item{filter_type}{The type of filter to apply in the ChEMBL API.}

\item{value}{The value to filter on in the ChEMBL API.}

\item{format}{The format of the response (default is "json").}
}
\value{
The response from the ChEMBL API in JSON format.
}
\description{
This function queries the ChEMBL API using the specified parameters and returns the response in JSON format.
}
\examples{
queryChemblAPI("mechanism", "molecule_chembl_id", "in", "CHEMBL1413")

}
