% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChem.R
\name{queryPubChem}
\alias{queryPubChem}
\title{Query the PubChem REST API, with the result automatically converted from
JSON to a list. This only works when \code{output='JSON'} in \code{getPubChem}.}
\usage{
queryPubChem(...)
}
\arguments{
\item{...}{Fallthrough arguments to \code{AnnotationGx::getPubChem} function.}
}
\description{
Query the PubChem REST API, with the result automatically converted from
JSON to a list. This only works when \code{output='JSON'} in \code{getPubChem}.
}
