% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChem.R
\name{queryRequestPubChem}
\alias{queryRequestPubChem}
\title{Query the PubChem REST API, with the result automatically converted from
JSON to a list. This only works when \code{output='JSON'} in \code{getRequestPubChem}.}
\usage{
queryRequestPubChem(..., query_only = FALSE)
}
\arguments{
\item{...}{Fallthrough arguments to \code{AnnotationGx::getRequestPubChem} function.}

\item{query_only}{}
}
\description{
Query the PubChem REST API, with the result automatically converted from
JSON to a list. This only works when \code{output='JSON'} in \code{getRequestPubChem}.
}
