% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniChem.R
\name{queryUniChem}
\alias{queryUniChem}
\title{Generic helper method to make arbitary REST API request to UniChem v2}
\usage{
queryUniChem(
  endpoint,
  ...,
  verb = "POST",
  url = "https://www.ebi.ac.uk/unichem/api/v1"
)
}
\arguments{
\item{endpoint}{\code{character(1)} The endpoint to query. Valid endpoints at
time of writing are "sources", "compounds" and "connectivity".
Appended to the end of \code{url} with "/".}

\item{...}{Fall through arguments to \code{httr::RETRY}.}

\item{verb}{\code{character(1)} HTTP verb to query the API with. Valid options
for UniChem are "GET" and "POST".}

\item{url}{\code{character(1)} UniChem API URL. Don't change unless you know what
you are doing.}
}
\value{
\code{httr::response} object with the result of your query.
}
\description{
Generic helper method to make arbitary REST API request to UniChem v2
}
