% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniProt.R
\name{queryUniProt}
\alias{queryUniProt}
\title{Query the UniProt website REST API}
\usage{
queryUniProt(
  query,
  columns = "",
  limit = "",
  offset = "",
  format = "xml",
  include = TRUE,
  compress = TRUE,
  ...,
  url = "https://www.uniprot.org/uniprot"
)
}
\arguments{
\item{query}{A \code{character} vector of the API query to return. See
https://www.uniprot.org/help/text-search for information on query
structure.}

\item{format}{A \code{character} vector specifying the return format. Options
are 'html', 'tab', 'xls', 'fasta', 'gff', 'txt', 'xml', 'rdf', 'list' or
'rss'. Default is 'xml'. See https://www.uniprot.org/help/api\%5Fqueries
for more information.}

\item{include}{A \code{boolean} indicating if isoforms or description of
referenced data should be included.}

\item{...}{Catch all unnamed arguments. All parameters after ... must be
specified with as name=value or they will be ignored.}

\item{url}{A \code{character} vector of the URL for the REST API. Should not
include trailing '/', this will be added inside the function. This
parameter must be passed named or it will not work.}
}
\value{
A httr::response object containing the query results as \code{format}.
}
\description{
Query the UniProt website REST API
}
