% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-webScraping.R
\name{scrapeRemoteFTPFilesTable}
\alias{scrapeRemoteFTPFilesTable}
\title{Retrieve a table of files available from an FTP download site by scraping
the first HTML table at the specified URL.}
\usage{
scrapeRemoteFTPFilesTable(url, column = "Name", recursive = FALSE)
}
\arguments{
\item{url}{\code{character(1)}}

\item{column}{\code{character(1)} Name of the column containing the FTP file paths.
Defaults to "Name".}

\item{recursive}{\code{logical(1)} Should remote directories be searched recursively?
Defaults to \code{FALSE}.}
}
\value{
\code{data.table} Table of available files and their download URLs and
any additional metadata available in the HTML table at \code{url}.
}
\description{
Retrieve a table of files available from an FTP download site by scraping
the first HTML table at the specified URL.
}
