% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellosaurusAPI.R
\name{searchCellosaurusAPI}
\alias{searchCellosaurusAPI}
\title{Search Cellosaurus API}
\usage{
searchCellosaurusAPI(
  query,
  from = "id",
  extResource = NULL,
  to = c("id", "ac", "ca", "sx", "ag", "di", "derived-from-site", "misspelling"),
  numResults = 1,
  returnURL = FALSE
)
}
\arguments{
\item{query}{The query string to search for in the Cellosaurus API.}

\item{from}{The field to search in. Default is "id".}

\item{extResource}{The external resource to search in. Default is NULL.}

\item{to}{The fields to return in the response. Default is "id".}

\item{numResults}{The number of results to return. Default is 1.}

\item{returnURL}{Logical indicating whether to return the constructed API URL. Default is FALSE.}
}
\description{
This function searches the Cellosaurus API for cell line information based on the provided query.
}
\examples{
searchCellosaurusAPI("MCF-7")

}
