% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-standardize_names.R
\name{standardize_names}
\alias{standardize_names}
\title{Standardize Names}
\usage{
standardize_names(object)
}
\arguments{
\item{object}{A character vector containing the names to be standardized.}
}
\value{
A character vector with the standardized names.
}
\description{
This function takes a character vector and standardizes the names by converting them to lowercase,
removing any trailing information after a comma, removing any information within square brackets or parentheses,
removing any non-alphanumeric characters, replacing empty names with "invalid", and converting the names to uppercase.
}
\examples{
standardize_names(c("John Doe", "Jane Smith (Manager)", "Alice, PhD"))
# Output: [1] "JOHNDOE" "JANESMITH" "ALICE"
}
