% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchNested-methods.R
\name{unlistNested}
\alias{unlistNested}
\title{Unlists a nested list and removes NA values and duplicates.}
\usage{
unlistNested(element)
}
\arguments{
\item{element}{The nested list to be unlisted.}
}
\value{
A vector with NA values and duplicates removed.
}
\description{
This function takes a nested list as input and unlists it recursively.
It then removes any NA values and duplicates from the resulting vector.
}
\examples{
nested_list <- list(a = list(1, 2, NA), b = list(3, 4, 5))
unlistNested(nested_list)
# Output: [1] 1 2 3 4 5

}
