% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{wIdentifierToInchiKey}
\alias{wIdentifierToInchiKey}
\title{Parallelized wrapper for identifierToInchiKey function}
\usage{
wIdentifierToInchiKey(chemical_ids, target_names)
}
\arguments{
\item{chemical_ids}{\code{character} vector which is the src_compound_id}

\item{target_names}{\code{character} vector which is the id for the database/source}
}
\value{
\code{character} vector with the inchikey and inchi structure
}
\description{
Wrapper function takes in vector of chemical ids and source ids and returns the
inchi and inchikey structure using parallel computation
}
