% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{wInchiToDatabaseID}
\alias{wInchiToDatabaseID}
\title{Parallelized wrapper for inchiToDatabaseID function}
\usage{
wInchiToDatabaseID(inchis, target_names)
}
\arguments{
\item{inchis}{A \code{character(1)} vector containing the InchiKey if \code{type}="key"}

\item{target_names}{A \code{character()} vector specifying the name of the target
database id to return. If missing, returns all}
}
\value{
A \code{data.frame} mapping from the inchi or inchikey to the chemical source ids
}
\description{
Wrapper function takes in vector of inchikeys and target names and
returns the appropriate chemical source id using parallel computation
}
