% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniChem.R
\name{wMapBetweenSources}
\alias{wMapBetweenSources}
\title{Parallelized wrapper for mapBetweenSources function}
\usage{
wMapBetweenSources(chemical_ids, src_name, target_name)
}
\arguments{
\item{chemical_ids}{\code{character} vector which is the compound identifier for
the specified database/source}

\item{src_name}{\code{character} vector which is the short name for the database/
source for which we know the chemical_id}

\item{target_name}{\code{character} vector which is the short name for the database/
source for which we require the compounds end_point}
}
\value{
A \code{character} vector which is the src_compound_id for the target_name
}
\description{
Wrapper function which returns the source compound ids for specific
database_id using parallel computation.
}
