% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodoAPI.R
\name{zenodoMetadata}
\alias{zenodoMetadata}
\alias{zenodo_metadata}
\title{Build the required metadata for uploading a file to zenodo}
\usage{
zenodoMetadata(
  title = "New Upload",
  upload_type = "dataset",
  description = "Default description",
  names = "Haibe-Kains, Benjamin",
  affiliations = list(c("Princess Margaret Cancer Centre")),
  ...
)
}
\arguments{
\item{title}{\code{character(1)} A character vector with length greater than 3.
Indicates the name of the Zenodo data entry for a new upload.}

\item{upload_type}{\code{character(1)} Type of Zenodo entry to create. Default
is "dataset".}

\item{description}{\code{character(1)} Body of the Description field on the
Zenodo entry.}

\item{names}{\code{character} One or more author names, formatted like
"Last Name, First Name". Defaults to "Haibe-Kains, Benjamin"}

\item{affiliations}{\code{list(1)} or \code{list(length(names))} List of character
vectors specifying author affiliations. May be length one, in which
case the character vector of affiliations is recycled for all authors, or
specify different vectors for each author name in \code{names}.}

\item{...}{Additional valid metadata items for Zenodo. No validation is done
yet, so make sure you know what you are doing before using this.}
}
\value{
\code{list} Named list with properly formatted metadata
}
\description{
Build the required metadata for uploading a file to zenodo
}
