% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationGx-package.R
\docType{package}
\name{AnnotationGx-package}
\alias{AnnotationGx}
\alias{AnnotationGx-package}
\title{AnnotationGx: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jermiah Joseph \email{jermiah.joseph@gmail.com}

}
\keyword{internal}
