% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_core.R
\name{AttributeSet}
\alias{AttributeSet}
\title{BioMart Attribute Set Class}
\description{
Represents a set of attributes from a BioMart dataset.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{attributes}}{List of AttributeInfo objects}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AttributeSet-new}{\code{AttributeSet$new()}}
\item \href{#method-AttributeSet-get_by_display_name}{\code{AttributeSet$get_by_display_name()}}
\item \href{#method-AttributeSet-print}{\code{AttributeSet$print()}}
\item \href{#method-AttributeSet-as.list}{\code{AttributeSet$as.list()}}
\item \href{#method-AttributeSet-filter}{\code{AttributeSet$filter()}}
\item \href{#method-AttributeSet-clone}{\code{AttributeSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttributeSet-new"></a>}}
\if{latex}{\out{\hypertarget{method-AttributeSet-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new AttributeSet object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeSet$new(attributes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{attributes}}{List of AttributeInfo objects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttributeSet-get_by_display_name"></a>}}
\if{latex}{\out{\hypertarget{method-AttributeSet-get_by_display_name}{}}}
\subsection{Method \code{get_by_display_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeSet$get_by_display_name(display_names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{display_names}}{Character vector of display names}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of matching AttributeInfo objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttributeSet-print"></a>}}
\if{latex}{\out{\hypertarget{method-AttributeSet-print}{}}}
\subsection{Method \code{print()}}{
Print method for AttributeSet objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeSet$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to print methods (not used)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttributeSet-as.list"></a>}}
\if{latex}{\out{\hypertarget{method-AttributeSet-as.list}{}}}
\subsection{Method \code{as.list()}}{
Convert AttributeSet to list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeSet$as.list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of attributes DisplayName as strings
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttributeSet-filter"></a>}}
\if{latex}{\out{\hypertarget{method-AttributeSet-filter}{}}}
\subsection{Method \code{filter()}}{
Filter attributes based on regex pattern
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeSet$filter(pattern, exclude = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pattern}}{Regular expression pattern to match against display names}

\item{\code{exclude}}{Logical, if TRUE excludes matching patterns, if FALSE includes them}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new AttributeSet with filtered attributes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttributeSet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AttributeSet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
