% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationGx-data.R
\docType{data}
\name{CCLE_sampleMetadata}
\alias{CCLE_sampleMetadata}
\title{CCLE_sampleMetadata is some preprocessed sample metadata from the CCLE dataset}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1461 rows and 4 columns.
}
\usage{
CCLE_sampleMetadata
}
\description{
CCLE_sampleMetadata is some preprocessed sample metadata from the CCLE dataset
}
\keyword{datasets}
