% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationGx-data.R
\docType{data}
\name{CCLE_treatmentMetadata}
\alias{CCLE_treatmentMetadata}
\title{CCLE_treatmentMetadata is some preprocessed treatment metadata from the CCLE dataset}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 24 rows and 1 columns.
}
\usage{
CCLE_treatmentMetadata
}
\description{
CCLE_treatmentMetadata is some preprocessed treatment metadata from the CCLE dataset
}
\keyword{datasets}
