% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChEMBL.R
\name{compoundQuery}
\alias{compoundQuery}
\title{Query the ChEMBL compound API end-point}
\usage{
compoundQuery(field, filter_type, value)
}
\description{
A specialised function for querying the compund_record resource
from the ChEMBL API\tabular{l}{
   Field \cr
   compound_key \cr
   compound_name \cr
   document_chembl_id \cr
   record_id \cr
   src_id \cr
}
}
