% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellosaurusAPI.R
\name{createQueryURLs}
\alias{createQueryURLs}
\title{Create a list of query URLS for Cellosaurus API}
\usage{
createQueryURLs(
  api,
  cl_names,
  format = "txt",
  num_results = 1,
  GETfxn = "search/cell-line?",
  fields
)
}
\arguments{
\item{api}{is the link to the API to build the URL. i.e "https://api.cellosaurus.org/"}

\item{cl_names}{is a list of the cell line names}

\item{format}{is the type of format to return from the API. Can be "txt" or "json"}

\item{num_results}{is the number of of items to return, DEFAULT=1}

\item{GETfxn}{is the function to use on the cellosaurus website. Currently only supports "search/cell-line?"}

\item{fields}{is a list of desired fields to include in the response}
}
\value{
A list of URLS
}
\description{
This function creates a queryURL for the cellosaurus API using a list of cell line names
}
\details{
Function to create a URL query for Cellosaurus to search for a cell-line using its name
An example call:   computedURLs <- createQueryURLs(api = "https://api.cellosaurus.org/", cl_names = \link{"22rv1", "Hela"}, fields = c("id", "ac"))
}
