% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationGx-data.R
\docType{data}
\name{CTRP_sampleMetadata}
\alias{CTRP_sampleMetadata}
\title{CTRP_sampleMetadata is some preprocessed sample metadata from the CTRP dataset}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1107 rows and 2 columns.
}
\usage{
CTRP_sampleMetadata
}
\description{
CTRP_sampleMetadata is some preprocessed sample metadata from the CTRP dataset
}
\keyword{datasets}
