% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationGx-data.R
\docType{data}
\name{CTRP_treatmentMetadata}
\alias{CTRP_treatmentMetadata}
\title{CTRP_treatmentMetadata is some preprocessed treatment metadata from the CTRP dataset}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 545 rows and 2 columns.
}
\usage{
CTRP_treatmentMetadata
}
\description{
CTRP_treatmentMetadata is some preprocessed treatment metadata from the CTRP dataset
}
\keyword{datasets}
