% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_core.R
\name{DatasetInfo}
\alias{DatasetInfo}
\title{BioMart Dataset Information Class}
\description{
Class that stores information about a dataset within a BioMart database.
}
\details{
This class encapsulates metadata about a specific dataset in a
BioMart database,
including its name, description, display name, and a reference
to the parent mart.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The internal name of the dataset}

\item{\code{description}}{A textual description of the dataset}

\item{\code{displayName}}{The human-readable name of the dataset}

\item{\code{mart}}{The MartInfo object this dataset belongs to}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DatasetInfo-new}{\code{DatasetInfo$new()}}
\item \href{#method-DatasetInfo-print}{\code{DatasetInfo$print()}}
\item \href{#method-DatasetInfo-clone}{\code{DatasetInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetInfo-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DatasetInfo object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetInfo$new(name, description, displayName, mart)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character, the internal name of the dataset}

\item{\code{description}}{Character, the dataset description}

\item{\code{displayName}}{Character, the human-readable name}

\item{\code{mart}}{MartInfo object, the parent mart}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new DatasetInfo object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetInfo-print"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetInfo-print}{}}}
\subsection{Method \code{print()}}{
Print method for DatasetInfo objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetInfo$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments (not used)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatasetInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DatasetInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
