% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChem-helpers.R
\name{.checkThrottlingStatus}
\alias{.checkThrottlingStatus}
\title{Checks to see if the PubChem query is exceeding the throttling limit}
\usage{
.checkThrottlingStatus(result, throttleMessage = FALSE)
}
\arguments{
\item{throttleMessage}{\code{logical} whether to print the throttling message}

\item{response}{\code{httr::response}}
}
\value{
\code{logical} whether the query is throttled
}
\description{
Checks to see if the PubChem query is exceeding the throttling limit
}
