% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChem-helpers.R
\name{.checkThrottlingStatus}
\alias{.checkThrottlingStatus}
\title{Checks the throttling status of a request and sleeps if necessary}
\usage{
.checkThrottlingStatus(result, throttleMessage = FALSE)
}
\arguments{
\item{result}{The result of the request}

\item{throttleMessage}{Logical indicating whether to print the throttling
message}
}
\value{
Logical indicating whether the request was throttled
}
\description{
This function checks the throttling status of a request by analyzing the
"x-throttling-control" header in the response. If the throttling status
indicates that the request is being throttled, the function sleeps for a
certain amount of time based on the percentage of throttling. The sleep time
is determined as follows:
\itemize{
\item If the percentage of throttling is greater than 50\%, the function sleeps
for 30 seconds.
\item If the percentage of throttling is greater than 30\%, the function sleeps
for 20 seconds.
\item If the percentage of throttling is greater than 15\%, the function sleeps
for 15 seconds.
\item Otherwise, the function sleeps for the same amount of time as the
percentage of throttling.
}
}
