% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChem.R
\name{.getPubChemAnnotationDT}
\alias{.getPubChemAnnotationDT}
\title{Function that returns a DT of getPubChemAnnotation results}
\usage{
.getPubChemAnnotationDT(compound, annotationType, ...)
}
\description{
Function that returns a DT of getPubChemAnnotation results
}
