% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChem.R
\name{.parseATCresponse}
\alias{.parseATCresponse}
\title{Function that parses the results of the PubChem PUG-VIEW API for the ATC Code header}
\usage{
.parseATCresponse(result)
}
\arguments{
\item{result}{The result of the PubChem PUG-VIEW API call.}
}
\value{
The ATC code as a string, or "NA" if no ATC code is found.
}
\description{
This function takes the result of the PubChem PUG-VIEW API and extracts the ATC Code header information.
It returns the ATC code as a string.
}
