% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChem.R
\name{.parseCASresponse}
\alias{.parseCASresponse}
\title{Function that parses the results of the PubChem PUG-VIEW API for the CAS header}
\usage{
.parseCASresponse(result)
}
\description{
Function that parses the results of the PubChem PUG-VIEW API for the CAS header
}
