% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChem.R
\name{.parseCHEMBLresponse}
\alias{.parseCHEMBLresponse}
\title{Function that parses the results of the PubChem PUG-VIEW API for the CHEMBL ID header}
\usage{
.parseCHEMBLresponse(result)
}
\description{
Function that parses the results of the PubChem PUG-VIEW API for the CHEMBL ID header
}
