% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_core.R
\name{FilterInfo}
\alias{FilterInfo}
\title{BioMart Filter Information Class}
\description{
Represents a filter from a BioMart dataset.
}
\details{
This class encapsulates information about a filter available in a BioMart dataset.
Filters are used to specify conditions for querying and subsetting data from BioMart.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The internal name of the filter used in BioMart queries}

\item{\code{displayName}}{The human-readable name of the filter shown in user interfaces}

\item{\code{description}}{A textual description explaining the filter's purpose}

\item{\code{type}}{The data type of the filter (e.g., "string", "boolean", "list")}

\item{\code{isHidden}}{Logical flag indicating if the filter should be hidden in user interfaces}

\item{\code{values}}{Possible values for the filter if it has a fixed/enumerated set of options}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FilterInfo-new}{\code{FilterInfo$new()}}
\item \href{#method-FilterInfo-print}{\code{FilterInfo$print()}}
\item \href{#method-FilterInfo-clone}{\code{FilterInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-FilterInfo-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new FilterInfo object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterInfo$new(
  name,
  displayName = NULL,
  description = NULL,
  type = NULL,
  isHidden = NULL,
  values = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character, the internal name of the filter used in API calls}

\item{\code{displayName}}{Character, the human-readable name shown to users}

\item{\code{description}}{Character, the filter description explaining its purpose}

\item{\code{type}}{Character, the data type of the filter (e.g., "string", "boolean")}

\item{\code{isHidden}}{Logical, whether the filter should be hidden in UIs}

\item{\code{values}}{List or vector, possible values for the filter if applicable}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new FilterInfo object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterInfo-print"></a>}}
\if{latex}{\out{\hypertarget{method-FilterInfo-print}{}}}
\subsection{Method \code{print()}}{
Print method for FilterInfo objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterInfo$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to print methods (not used)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns self for method chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FilterInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
