% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gCSI_sampleMetadata}
\alias{gCSI_sampleMetadata}
\title{gCSI_sampleMetadata is some preprocessed sample metadata from the NCI60 dataset}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1350 rows and 2 columns.
}
\usage{
gCSI_sampleMetadata
}
\description{
gCSI_sampleMetadata is some preprocessed sample metadata from the NCI60 dataset
}
\keyword{datasets}
