% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationGx-data.R
\docType{data}
\name{gCSI_treatmentMetadata}
\alias{gCSI_treatmentMetadata}
\title{gCSI_treatmentMetadata is some preprocessed treatment metadata from the gCSI dataset}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 44 rows and 2 columns.
}
\usage{
gCSI_treatmentMetadata
}
\description{
gCSI_treatmentMetadata is some preprocessed treatment metadata from the gCSI dataset
}
\keyword{datasets}
