% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationGx-data.R
\docType{data}
\name{GDSC_sampleMetadata}
\alias{GDSC_sampleMetadata}
\title{gdsc_sampleMetadata is some preprocessed sample metadata from the GDSC dataset}
\format{
A data table with 5 columns and 1001 rows.
\describe{
\item{GDSC.Sample_Name}{\code{char} The name of the cell line in the GDSC dataset.}
\item{GDSC.COSMIC_ID}{\code{int} The COSMIC ID of the cell line in the GDSC dataset.}
}
}
\source{
https://www.cancerrxgene.org/
}
\usage{
data(GDSC_sampleMetadata)
}
\description{
A preprocessed version of the sample metadata from the GDSC dataset. This dataset
contains the following columns: GDSC.Sample_Name, GDSC.BROAD_ID, GDSC.RRID, GDSC.COSMIC_ID, and CCLE.sampleid.
This dataset is used in the AnnotationGx package to map cell line names from various sources to the
Cellosaurus database.
}
\examples{
data(GDSC_sampleMetadata)
head(GDSC_sampleMetadata)
}
\keyword{datasets}
