% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GDSC_treatmentMetadata}
\alias{GDSC_treatmentMetadata}
\title{GDSC_treatmentMetadata is some preprocessed treatment metadata from the GDSC dataset}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 621 rows and 3 columns.
}
\usage{
GDSC_treatmentMetadata
}
\description{
GDSC_treatmentMetadata is some preprocessed treatment metadata from the GDSC dataset
}
\keyword{datasets}
