% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellosaurus.R
\name{getCellosaurus}
\alias{getCellosaurus}
\title{Build a \code{data.table} of annotations from query list.}
\usage{
getCellosaurus(
  cellnames,
  url = "https://ftp.expasy.org/databases/cellosaurus/cellosaurus.xml"
)
}
\arguments{
\item{cellnames}{is the list of cell names (identifiers and synonyms)}

\item{url}{is cellosaurus link to xml}
}
\description{
Build a \code{data.table} of annotations from query list.
}
