% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellosaurus.R
\name{getCelloxml}
\alias{getCelloxml}
\title{Access and parse Cellosaurus xml for regular users}
\usage{
getCelloxml(
  url = "https://ftp.expasy.org/databases/cellosaurus/cellosaurus.xml",
  verbose = TRUE
)
}
\arguments{
\item{url}{is cellosaurus link to xml. This should be a valid link to Cellosaurus xml. Default is \verb{https://ftp.expasy.org/databases/cellosaurus/cellosaurus.xml}.}

\item{verbose}{is TRUE by default}
}
\value{
An \code{XML} document of Cellosaurus
}
\description{
This function reads Cellosaurus XML and parses nodes for cell line annotations based on input parameters, returning a data.table object
}
\details{

}
\references{
Bairoch A.The Cellosaurus, a cell line knowledge resource.J. Biomol. Tech. 29:25-38(2018) DOI: 10.7171/jbt.18-2902-002; PMID: 29805321; PMCID: PMC5945021
}
