% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellosaurus.R
\name{getCelloxml}
\alias{getCelloxml}
\title{Access and parse Cellosaurus xml}
\usage{
getCelloxml(
  url = "https://ftp.expasy.org/databases/cellosaurus/cellosaurus.xml"
)
}
\arguments{
\item{url}{is cellosaurus link to xml}
}
\value{
A data.table object with the results of the input cell line(s)
}
\description{
This function reads Cellosaurus XML and parses nodes for cell line annotations based on input parameters, returning a data.table object
}
\details{

}
\references{
Bairoch A.The Cellosaurus, a cell line knowledge resource.J. Biomol. Tech. 29:25-38(2018) DOI: 10.7171/jbt.18-2902-002; PMID: 29805321; PMCID: PMC5945021
}
