% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChEMBL.R
\name{getChemblMechanism}
\alias{getChemblMechanism}
\title{Query the ChEMBL mechanism API end-point for a given CHEMBL_ID or list of CHEMBL_IDs}
\usage{
getChemblMechanism(
  chembl.ID,
  resources = "mechanism",
  field = "molecule_chembl_id",
  filter_type = "in",
  returnURL = FALSE
)
}
\arguments{
\item{chembl.ID}{\code{character(1)} or \code{character(n)} CHEMBL_ID(s) to query}

\item{resources}{\code{character(1)} Resource to query. Default is "mechanism"}

\item{field}{\code{character(1)} or \code{character(n)} Field(s) to query. Default is "molecule_chembl_id"}

\item{filter_type}{\code{character(1)} Filter type. Default is "in"}

\item{returnURL}{\code{logical(1)} Return the URL(s) instead of the data table}
}
\value{
\code{data.table} A table containing all mechanism of action entries and the CHEMBL_IDs queried
}
\description{
A specialised function for querying the mechanism resource from the ChEMBL API.
}
\details{
Query the ChEMBL mechanism API end-point for a given CHEMBL_ID or list of CHEMBL_IDs
}
\examples{
getChemblMechanism("CHEMBL1413")

chembl_ids <- c("CHEMBL515", "CHEMBL235191", "CHEMBL1413")
getChemblMechanism(chembl_ids)

}
