% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGencode.R
\name{getGencodeFile}
\alias{getGencodeFile}
\title{Download files from the Gencode FTP site and load them as the appropriate
Bioconductor classes.}
\usage{
getGencodeFile(
  file = "gencode\\\\.{v}\\\\.annotation\\\\.gtf\\\\.gz",
  type = infer_gencode_type(file),
  version = "latest",
  chr = c("GRCh38", "GRCh37"),
  dir = tempdir(),
  url = "https://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human"
)
}
\arguments{
\item{file}{\code{character(1)} String name of file to download from the
Gencode FTP site. See \code{getGencodeAvailableFiles()} for options. Supports
regex to match file names. If your version is specified as "{v}" it will
be interpolated from the \code{version} argument.}

\item{type}{\code{character(1)} One of "GTF", "GFF3", "FASTA" or "metadata".
Defaults to \code{infer_gencode_type(file)}, which guesses based on the
\code{file} string.}

\item{version}{\code{character(1)} Gencode version to download from.
Defaults to "latest". See \code{?getGencodeFTPTable} for options.}

\item{dir}{\code{character(1)} Path to download the file to. Defaults to
\code{tempdir()}. When this value is \code{tempdir()} the downloaded file is
automatically deleted when the function exits.}

\item{url}{\code{character(1)} Address of Gencode FTP web page. Default is page
for Gencode Human files.}
}
\value{
\code{GenomicRanges} object when type="GTF", \code{DNAStringSet} when
type="FASTA", or \code{data.table}/\code{character} (as appropriate) when
type="metadata".
}
\description{
File descriptions are available at https://www.gencodegenes.org/human/.
}
\seealso{
Other Gencode: 
\code{\link{getGencodeAvailableFiles}()},
\code{\link{getGencodeFTPTable}()},
\code{\link{getGencodeFilesTable}()},
\code{\link{getGencodeGRangesAnnotated}()},
\code{\link{infer_gencode_type}()}
}
\concept{Gencode}
