% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncotree.R
\name{getOncotreeMainTypes}
\alias{getOncotreeMainTypes}
\title{Get the main types from the Oncotree database.}
\usage{
getOncotreeMainTypes()
}
\value{
A \code{data.table} containing the main types from the Oncotree database.
}
\description{
This function retrieves the main types from the Oncotree database.
}
