% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncotree.R
\name{getOncotreeTumorTypes}
\alias{getOncotreeTumorTypes}
\title{Get the tumor types from the Oncotree database.}
\usage{
getOncotreeTumorTypes()
}
\value{
A \code{data.table} containing the tumor types from the Oncotree database.
}
\description{
This function retrieves the tumor types from the Oncotree database.
}
