% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChem.R
\name{getPubChemAnnotations}
\alias{getPubChemAnnotations}
\title{Retrieve PubChem annotations for a given compound}
\usage{
getPubChemAnnotations(compound, annotations, ...)
}
\arguments{
\item{compound}{The compound for which PubChem annotations are to be retrieved.}

\item{annotations}{A character vector specifying the annotations to retrieve.}

\item{...}{Additional arguments to be passed to getPubChemAnnotation().}
}
\value{
A merged data table containing the PubChem annotations for the specified compound.
}
\description{
This function retrieves PubChem annotations for a given compound using the specified annotations.
}
\examples{
getPubChemAnnotations(
     compound = "36314", 
     annotations= c('ChEMBL ID', 'NSC Number', 'Drug Induced Liver Injury'))

}
