% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem_rest.R
\name{getPubchemCompound}
\alias{getPubchemCompound}
\title{Retrieve PubChem compound information}
\usage{
getPubchemCompound(
  ids,
  from = "cid",
  to = "property",
  properties = c("Title", "InChIKey"),
  raw = FALSE,
  query_only = FALSE,
  verbose = FALSE,
  output = "JSON",
  ...
)
}
\arguments{
\item{ids}{A vector of compound identifiers.}

\item{from}{The source namespace of the compound identifiers. Default is 'cid'.}

\item{to}{The target namespace for the compound information. Default is 'property'.}

\item{properties}{A character vector specifying the properties to retrieve.}

\item{raw}{Logical indicating whether to return the raw query results. Default is FALSE.}

\item{query_only}{Logical indicating whether to only perform the query without retrieving the results. Default is FALSE.}

\item{verbose}{Logical indicating whether to display verbose output. Default is FALSE.}

\item{output}{The format of the query results. Default is 'JSON'.}

\item{...}{Additional arguments to be passed to the query_pubchem_rest function.}
}
\value{
A data.table containing the retrieved compound information.
}
\description{
This function retrieves compound information from PubChem using the PubChem REST API.
}
\examples{
properties=c('Title', 'MolecularFormula', 'InChIKey', 'CanonicalSMILES')
getPubchemCompound(c(3672, 176870), from = 'cid', to = 'property', properties = properties)

}
