% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChem.R
\name{getPubChemFromNSC}
\alias{getPubChemFromNSC}
\title{getPubChemFromNSC}
\usage{
getPubChemFromNSC(
  ids,
  to = "cids",
  ...,
  batch = TRUE,
  raw = FALSE,
  options = NA,
  query_only = FALSE
)
}
\arguments{
\item{ids}{A \code{character} or \code{numeric} vector of valid NSC ids to use for the
query.}

\item{to}{A \code{character(1)} vector with the desired return type. Currently
only 'cids' and 'sids' are implemented, but other options are available
via the PubChem API. This corresponds to the \code{operation} portion of the
PubChem API URL Path.}

\item{...}{Fall through arguments to bpmapply. Use this to pass in BPPARAM
parameter to customize parellization settings. Alternatively, just call
\code{register()} with your desired parallel backend configuration.}

\item{raw}{A \code{logical(1)} vector specifying whether to early return the raw
query results. Use this if specifying an unimplemented return to the \code{to}
parameter.}
}
\value{
A \code{data.table} where the first column is the specified NSC ids and
the second column is the results specified in \code{to}.
}
\description{
Return a data.table mapping from ids to the information specified in \code{to}.
}
