% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem_rest.R
\name{getPubchemProperties}
\alias{getPubchemProperties}
\title{Retrieves the PubChem XML schema and extracts property information.}
\usage{
getPubchemProperties()
}
\value{
A data table containing the extracted property information.
}
\description{
This function retrieves the PubChem XML schema from the specified URL and
extracts the property information from it. The property information includes
the name and type of each property.
}
