% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChem-helpers.R
\name{getPubChemStatus}
\alias{getPubChemStatus}
\title{simple checkThrottling Status by a default response}
\usage{
getPubChemStatus(sleep = FALSE)
}
\arguments{
\item{sleep}{A logical value indicating whether to sleep before checking
the throttling status.}
}
\value{
A logical value indicating whether the query is throttled.
}
\description{
This function checks the throttling status of a PubChem query and returns
a logical value indicating whether the query is throttled.
}
\examples{
getPubChemStatus()
getPubChemStatus(sleep = TRUE)

}
\keyword{internal}
