% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem_status.R
\name{getPubchemStatus}
\alias{getPubchemStatus}
\title{Retrieves the status of a PubChem request}
\usage{
getPubchemStatus(
  returnMessage = FALSE,
  printMessage = TRUE,
  url = "https://pubchem.ncbi.nlm.nih.gov/rest/pug/compound/name/Aspirin/cids/JSON"
)
}
\arguments{
\item{returnMessage}{Logical indicating whether to return the parsed information from the response.}

\item{printMessage}{Logical indicating whether to print the status message.}

\item{url}{The URL to send the request to. Default is "https://pubchem.ncbi.nlm.nih.gov/rest/pug/compound/name/Aspirin/cids/JSON".}
}
\value{
The status code of the response. If \code{returnMessage} is \code{TRUE}, the parsed information from the response is also returned.
}
\description{
This function sends a request to PubChem to retrieve the status of a given URL.
It returns the status code and, if specified, the parsed information from the response.
}
\examples{
getPubchemStatus()
getPubchemStatus(returnMessage = TRUE)
getPubchemStatus(printMessage = FALSE)

}
