% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChem.R
\name{getPubChemSubstance}
\alias{getPubChemSubstance}
\title{Retrieve PubChem substance information}
\usage{
getPubChemSubstance(
  ids,
  from = "cid",
  to = "sids",
  ...,
  batch = TRUE,
  raw = FALSE
)
}
\arguments{
\item{ids}{A character vector of IDs to query.}

\item{from}{The namespace of the input IDs (default is 'cid').}

\item{to}{The namespace to map the input IDs to (default is 'sids').}

\item{...}{Additional arguments to be passed to the queryPubChem function.}

\item{batch}{Logical indicating whether to use batch queries (default is TRUE).}

\item{raw}{Logical indicating whether to return the raw query results (default is FALSE).}
}
\value{
A data table containing the retrieved PubChem substance information.
If raw is TRUE, the function returns the raw query results.
If any queries fail, the function includes the failed queries in the returned object.
}
\description{
This function retrieves PubChem substance information based on the provided IDs.
It supports mapping from various namespaces such as CID, name, xref, smiles, inchi, and sdf.
The function can handle batch queries and can return the raw query results if specified.
}
\references{
Documentation for the PubChem API: https://pubchemdocs.ncbi.nlm.nih.gov/
}
