% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unichem.R
\name{getUnichemSources}
\alias{getUnichemSources}
\title{Get the list of sources in UniChem.}
\usage{
getUnichemSources(all_columns = FALSE)
}
\arguments{
\item{all_columns}{\code{boolean} Whether to return all columns. Defaults to FALSE.

Returns a \code{data.table} with the following columns:
\itemize{
\item \code{CompoundCount} (integer): Total of compounds provided by that source
\item \code{BaseURL} (string): Source Base URL for compounds
\item \code{Description} (string): Source database description
\item \code{LastUpdated} (string): Date in which the source database was last updated
\item \code{Name} (string): Short name of the source database
\item \code{NameLabel} (string): Machine readable label name of the source database
\item \code{NameLong} (string): Full name of the source database
\item \code{SourceID} (integer): Unique ID for the source database
\item \code{Details} (string): Notes about the source
\item \code{ReleaseDate} (string): Date in which the source database was released
\item \code{ReleaseNumber} (integer): Release number of the source database data stored in UniChEM
\item \code{URL} (string): Main URL for the source
\item \code{UpdateComments} (string): Notes about the update process of that source to UniChEM
}}
}
\value{
A data.table with the list of sources in UniChem.
}
\description{
Get the list of sources in UniChem.
}
