% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniChem.R
\name{getUniChemSources}
\alias{getUniChemSources}
\title{Fetch a table of cheminformatic databases which can be mapped between using
the UniChem 2.0 API}
\usage{
getUniChemSources(metadata = FALSE, ...)
}
\arguments{
\item{metadata}{\code{logical(1)} Should all the columns of the table be returned?
Default is \code{FALSE}, which selects only the database name ("name") and
UniChem source identifier ("sourceID") columns. When \code{TRUE} includes all
available source metadata, including long form descriptions of the databases,
their respective URLs, and much more.}

\item{...}{\code{pairlist} Fall through parameters to \code{httr::GET} via \code{httr:RETRY}.
Pass \code{httr::verbose()} to see full details of the query being constructed.}
}
\value{
\code{data.table} Table of UniChem source databases which can be mapped
between. The "name" column is the database name, and the "sourceID" column
is its unique integer ID in UniChem. When \code{metadata=TRUE} also includes
additional source database metadata.
}
\description{
Fetch a table of cheminformatic databases which can be mapped between using
the UniChem 2.0 API
}
\details{
For more details see the Unichem 2.0 API documentation at:
https://chembl.gitbook.io/unichem/unichem-2.0/unichem-2.0-beta/whats-new
}
\author{
Christopher Eeles (christopher.eeles@uhnresearch.ca)
}
