% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{mapCell2Accession}
\alias{mapCell2Accession}
\title{Maps cell line names to accession numbers}
\usage{
mapCell2Accession(names, numResults = 1, from = "id", ...)
}
\arguments{
\item{names}{A vector of cell line names to be mapped.}

\item{numResults}{The maximum number of results to retrieve for each cell line name.}

\item{from}{The field to search for the cell line name in the Cellosaurus database.}

\item{...}{Additional arguments to be passed to the Cellosaurus API.}
}
\value{
A data table with the mapping information, including the cell line name,
accession number, and Cellosaurus ID.
}
\description{
This function takes a vector of cell line names and maps them to accession numbers
using the Cellosaurus database. It performs a parallel request to retrieve the
mapping information and returns a data table with the results.
}
\examples{
mapCellline2Accession(c("A549", "HeLa"))
}
