% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{mapCell2Accession}
\alias{mapCell2Accession}
\title{Maps cell line names to accession numbers}
\usage{
mapCell2Accession(
  ids,
  numResults = 1,
  from = "id",
  to = c("id", "ac"),
  query_only = FALSE,
  raw = FALSE,
  ...
)
}
\arguments{
\item{ids}{A character vector of cell line names.}

\item{numResults}{The number of results to return for each query. Default is 1.}

\item{from}{The field to query from. Default is "id".}

\item{to}{The field to query to. Default is both "id" and "ac".}

\item{query_only}{If TRUE, returns the query URL instead of the results. Default is FALSE.}

\item{raw}{If TRUE, returns the raw response instead of a data table. Default is FALSE.}

\item{...}{Additional arguments to pass to the request.}
}
\value{
A data table with the mapping information, including the cell line name,
accession number, and Cellosaurus ID.
}
\description{
This function takes a vector of cell line names and maps them to accession numbers
using the Cellosaurus database. It performs a parallel request to retrieve the
mapping information and returns a data table with the results.
}
\examples{
mapCell2Accession(c("A549", "HeLa"))
}
