% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{mapCell2Accession}
\alias{mapCell2Accession}
\title{Maps cell line names to accession numbers}
\usage{
mapCell2Accession(
  ids,
  numResults = 1000,
  from = "id",
  to = c("id", "ac"),
  prioritizeParent = FALSE,
  query_only = FALSE,
  raw = FALSE,
  BPPARAM = BiocParallel::SerialParam(),
  ...
)
}
\arguments{
\item{ids}{A character vector of cell line names.}

\item{numResults}{The number of results to return for each query. Default is 1.}

\item{from}{The field to query from. Default is "id".}

\item{to}{The field to query to. Default is both "id" and "ac".}

\item{prioritizeParent}{If TRUE, prioritizes the parent cell line when multiple matches are found. Default is FALSE.
When prioritizeParent is TRUE, if multiple matches are found for a cell line name, the function
will prioritize the parent cell lines over other matches. This can be useful when dealing with
cell line hierarchies where the parent cell line represents a broader category.
An example of this is trying to map using id "BT474" which returns "CVCL_YX79" which corresponds
to "BT474 A3" whereas "BT-474" exists in the database as "CVCL_0179". If prioritizeParent is TRUE,
the function will prioritize "CVCL_0179" over "CVCL_YX79" since "BT-474" is the parent cell line of
"BT474 A3".}

\item{query_only}{If TRUE, returns the query URL instead of the results. Default is FALSE.}

\item{raw}{If TRUE, returns the raw response instead of a data table. Default is FALSE.}

\item{BPPARAM}{A BiocParallel parameter object controlling the parallelization.}

\item{...}{Additional arguments to pass to the request.}
}
\value{
Depending on parameters, either a:
\code{data.table} with the "id", "ac", "query", and "query:id" columns.
\code{list} of query URLs if \code{query_only} is TRUE.
\code{list} of raw responses if \code{raw} is TRUE.
}
\description{
This function takes a vector of cell line names and maps them to accession numbers
using the Cellosaurus database. It performs a parallel request to retrieve the
mapping information and returns a data table with the results.
}
\examples{
mapCell2Accession(c("A549", "HeLa"))
}
