% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem_rest.R
\name{mapCompound2CID}
\alias{mapCompound2CID}
\title{Map compound names to PubChem CIDs}
\usage{
mapCompound2CID(names, first = FALSE, ...)
}
\arguments{
\item{names}{A character vector of compound names.}

\item{first}{Logical indicating whether to return only the first CID for each compound name (default is FALSE).}

\item{...}{Additional arguments to be passed to the getPubchemCompound function.}
}
\value{
A character vector of PubChem CIDs.
}
\description{
This function maps compound names to PubChem CIDs using the PubChem REST API.
}
\examples{
mapCompound2CID(c("aspirin", "caffeine"))

}
