% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChem.R
\name{queryPubChem}
\alias{queryPubChem}
\title{queryPubChem}
\usage{
queryPubChem(
  id,
  domain = "compound",
  namespace = "cid",
  operation = NA,
  output = "JSON",
  ...,
  url = "https://pubchem.ncbi.nlm.nih.gov/rest/pug",
  operation_options = NA,
  batch = TRUE,
  raw = FALSE
)
}
\arguments{
\item{id}{The identifier for REST API query. This should be a valid compound
identifier for the specified \code{domain} and \code{namespace}. For example, if
\code{domain='compound'} and \code{namespace='cid'} then \code{id} should be one or
more compound cids to retrieve information about. If \code{id} is a \code{list} or
\code{vector}, the ids are parsed to a string with ids separated by
commas. This parameter forms the '<identifiers>' part of the Input
portion of the URL path: '<domain>/<namespace>/<identifiers>'.}

\item{domain}{Which records in PubChem are the subject of the query? Default
is 'compound'. Usually, this will be one of 'substance', 'compound', or
'assay'. For more advanced options see the PubChem PUG REST API
documentation. This forms the '<domain>' part of the Input portion of the
URL path.}

\item{namespace}{What kind of identifiers are in \code{id}? These options
are specific to the selected domain. For the 'compound' domain, common
options are 'cid', 'name', 'smiles', 'inchi', or 'formula'. For the
'substance' domain common options are 'sid' or 'name'. For the
'assay' domain common options are 'aid' and 'target/<assay_target>' where
'<assay_target>' is one of 'gi', 'proteinname', 'geneid', 'genesymbol' or
'accession'. For all options please see the PubChem PUG REST API
documentation. This parameter makes up the '<namespace>' part of the
Input portion of the URL path.}

\item{operation}{What kind of data to return for the specified input
portion of the URL path? If excluded this defaults to return the entire
record for that specified input. Options are domain specific and details
are available in the PubChem PUG REST API documentation. Options valid for
'compound', 'substance' and 'assay' domains are 'record', 'aids', 'sids',
'aids', 'description' and 'classification'. Common options for the
'compound' domain are 'assaysummary', 'conformers' and
'property/<properties>' where '<properties>' is a comma separated list
property tags; valid tags can be found at
https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest$_Toc494865565.
Common options for the 'substance' domain are 'assaysummary'.
Common options for the 'assay' domain include 'concise', 'summary' and
'targets/<target type>' where '<target type>' is one of 'ProteinGI',
'ProteinName', 'GeneID' or 'GeneSymbol'.}

\item{output}{What format should the data be returned in? Default is 'JSON'.
other common options include 'XML', 'CSV' and 'TXT'. For a complete list of
output format options, plase see the PubChem PUG REST API documentation.}

\item{...}{Fall through parameters to \code{bpmapply}.}

\item{url}{The URL of the PubChem REST API. Probably don't change this.}

\item{operation_options}{Further optional arguments for the selected operation.
this is specific to the selected operation. This is appended as a string
after '?' at the end of the query. See
https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest$_Toc494865565 for details.}
}
\description{
queryPubChem
}
