% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniChem.R
\name{queryUniChemCompound}
\alias{queryUniChemCompound}
\title{Query the UniChem 2.0 compounds endpoint using POST requests}
\usage{
queryUniChemCompound(...)
}
\arguments{
\item{...}{\code{pairlist} Fall through parameters to \code{httr::POST} via
\code{httr:RETRY}. Pass \code{httr::verbose()} to see full details of the query being
constructed.}

\item{url}{\code{character(1)} URL to query. Don't change this unless you know
what you are doing!}
}
\description{
Query the UniChem 2.0 compounds endpoint using POST requests
}
\author{
Christopher Eeles (christopher.eeles@uhnresearch.ca)
}
