% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unichem.R
\name{queryUnichemCompound}
\alias{queryUnichemCompound}
\title{Query UniChem for a compound.}
\usage{
queryUnichemCompound(
  compound,
  type,
  sourceID = NA_integer_,
  request_only = FALSE,
  raw = FALSE,
  ...
)
}
\arguments{
\item{compound}{\code{character} or \code{integer} The compound identifier to search for.}

\item{type}{\code{character} The type of compound identifier to search for. Valid types are "uci", "inchi", "inchikey", and "sourceID".}

\item{sourceID}{\code{integer} The source ID to search for if the type is "sourceID". Defaults to NULL.}

\item{request_only}{\code{boolean} Whether to return the request only. Defaults to FALSE.}

\item{raw}{\code{boolean} Whether to return the raw response. Defaults to FALSE.}

\item{...}{Additional arguments.}
}
\value{
A list with the external mappings and the UniChem mappings.
}
\description{
This function queries the UniChem API for a compound based on the provided parameters.
}
\examples{
queryUnichemCompound(type = "sourceID", compound = "444795", sourceID = 22)

}
