% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniChem.R
\name{queryUniChemCompounds}
\alias{queryUniChemCompounds}
\title{Query the UniChem 2.0 compounds endpoint using POST requests}
\usage{
queryUniChemCompounds(...)
}
\arguments{
\item{...}{\code{pairlist} Fall through parameters to \code{httr::POST} via
\code{httr:RETRY}. Pass \code{httr::verbose()} to see full details of the query being
constructed.}
}
\description{
Query the UniChem 2.0 compounds endpoint using POST requests
}
\author{
Christopher Eeles (christopher.eeles@uhnresearch.ca)
}
