% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-split.R
\name{strSplit}
\alias{strSplit}
\title{Split a character vector into a matrix based on a delimiter}
\usage{
strSplit(x, split, fixed = TRUE, n = Inf)
}
\arguments{
\item{x}{A character vector to be split}

\item{split}{A character string specifying the delimiter}

\item{fixed}{A logical value indicating whether the delimiter should be treated as a fixed string}

\item{n}{An integer specifying the maximum number of splits to be performed}
}
\value{
A matrix where each row represents a split element
}
\description{
This function splits a character vector into a matrix based on a specified delimiter.
It can handle both finite and infinite splits.
}
\examples{
strSplit("Hello,World", ",")
# Output:
#      [,1]    [,2]   
# [1,] "Hello" "World"

}
